### 2023-12-14 #6149 @dfm
@annatartaglia — Thanks for your submission! Given the small size of this code base, I have submitted this pre-review for editorial review with respect to the ["substantial scholarly effort"](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) submission requirement. This usually takes a week or two and I'll report back here as soon as I hear back from the rest of the editorial board.
