### 2024-01-15 dfeb029
CompatHelper: bump compat for VLBISkyModels to 0.4, (keep existing compat)
