+++ Project.toml
@@ -28,7 +28,7 @@ NonlinearSolve = "1, 3"
 QuadGK = "2"
 SpecialFunctions = "2"
 StationaryRandomFields = "0.1"
-VLBISkyModels = "0.2, 0.3"
+VLBISkyModels = "0.2, 0.3, 0.4"
 julia = "1.9"
 
 [extras]
