### 2024-02-01 #6149 @kazuakiyama
@dfm Thank you very much for sharing this update. As a co-author of this submission, I have a question as this paper is our first experience working with JOSS. We have kept the package somewhat frozen since the submission to JOSS because potential reviewers may want to see the version of the submission. As the review may take longer, I would like to ask if we are allowed to keep the package continuously developed and updated. At the time of the submission, the package already had a full function of simulating scattering effects enough to work as a standalone package but we want to extend their functionalities.
