### 2024-02-09 #6149 @dfm
👋 @tomkimpson, @Edenhofer, @kiranshila — Would any of you be available and willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
