### 2024-02-14 #6354 @editorialbot
[REVIEW]: ScatteringOptics.jl: An Interstellar Scattering Framework in the Julia Programming Language
**Submitting author:** @annatartaglia (<a href="http://orcid.org/0000-0001-9229-8833">Anna Tartaglia</a>)
**Repository:** https://github.com/EHTJulia/ScatteringOptics.jl
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.9
