### 2024-02-16 #6354 @Edenhofer
I would like to request a COI waiver for my review of ScatteringOptics.jl (#6354) regarding Paul Tiede, who is a co-author of said paper. I've been at the Center for Astrophysics | Harvard & Smithsonian, Paul's employer, for about half a year in 2023 and have had two extended conversations with Paul about Gaussian processes, a subject unrelated to the submission. Since we work on very different topics in astrophysics, have never collaborated, and haven't interacted regularly, I believe I can evaluate the submission impartially.
