### 2024-02-16 #6354 @Edenhofer
## Paper

The paper is well written, summarizes the need for the software well, and focuses on the essential mathematics. However, as a non-expert in the field of radio astronomy, I think that the mathematics discussed should be more closely tied to what ScatteringOptics.jl provides. In particular, it is not clear to me how the equation after l70 and the equation after l74 are related. I probably am missing something, but I thought the scattering is performed in image space and only $G$ and $\phi$ are relevant for ScatteringOptics.jl. If ScatteringOptics.jl only returns the effective scattering in visual space, please state this more clearly, e.g. in l82ff.

### Minor notes

* l65 Why does $\phi$ depend twice on $r$?
* l65-l67 What is the difference between the phase screen $\phi_r$ and the spatial structure function of the phase screen $D_\phi(r)$?
* l70 How is $G$ related to $\phi$?
* r$ is defined in l71 and again in l65; please merge the definitions.
* l76 It is not clear to me what "earth-screen distance D" refers to; please rephrase.
* Is eq. following l74 an alternative formulation of l70 in visibility space?

* l97 It would be great if the authors could provide an example that is reproducible  from scratch, e.g., by providing "jason_mad_eofn.fits".
* l102 Wrap line.
* Figure 1 and Figure 2: Increase DPI for publication?
