### 2024-02-16 EHTJulia/ScatteringOptics.jl#22 @Edenhofer
Question regarding `kzetafinder_equation`
https://github.com/EHTJulia/ScatteringOptics.jl/blob/c730f8e9dfefb1656e17e0cd6970df1bc06cfd9e/src/kzetafinders/periodicboxcar.jl#L18-L21

The referenced formula (Equation 47 of Psaltis et al. 2018) contains an additional `kzeta` in the argument to sinus. Is that potentially missing here?

Part of openjournals/joss-reviews#6354 .
