### 2024-02-16 EHTJulia/ScatteringOptics.jl#23 @Edenhofer
Tests
I think it would be great if the repository would include some basic tests of the software as well as, e.g., an end-to-end test of a demonstration script.

Part of openjournals/joss-reviews#6354 .
