### 2024-02-16 EHTJulia/ScatteringOptics.jl#22 @Edenhofer
A related question about `kzetafinder_equation` in src/kzetafinders/dipole.jl and src/kzetafinders/vonmises.jl , shouldn't the square difference be penalized instead of `value - A^2`?
