### 2024-02-16 #6354 @dfm
> I would like to request a COI waiver for my review of ScatteringOptics.jl (https://github.com/openjournals/joss-reviews/issues/6354) regarding Paul Tiede, who is a co-author of said paper. I've been at the Center for Astrophysics | Harvard & Smithsonian, Paul's employer, for about half a year in 2023 and have had two extended conversations with Paul about Gaussian processes, a subject unrelated to the submission. Since we work on very different topics in astrophysics, have never collaborated, and haven't interacted regularly, I believe I can evaluate the submission impartially.

@Edenhofer — Thanks for bringing this up! Given the weak nature of this potential COI, I'm happy for you to continue with the review, having noted this context.

@annatartaglia — If you or any of your co-authors have any concerns about this at all please reach out to me here or over email (my address should be easy to find!). Thanks all!!
