### 2024-02-29 #6354 @tomkimpson
The package is useful and the accompanying paper is well-written. It was nice to learn about the use of Julia by the EHT collaboration. I have some general comments below. I will review the code itself separately and open any issues as needed.

* In the manuscript, is it really true that current scattering models are written in Python? If so, then it is fine to refer to the speed of Julia as the motivation. If not (I always thought these things were done in C++, but I could just be ignorant), then perhaps the composability with the wider EHTJulia organisation is a more accurate justification.
* In the manuscript, it claims that  `ScatteringOptics.jl includes a set of abstract types that enable users to define other phase screen models.` It is not obvious to me how a user would define some other phase screen model - perhaps a MWE would be useful.
* The MWE in the tutorial is difficult for a user to reproduce without the accompanying `.fits` file. If sufficiently small, it could be worth including `jason_mad_eofn.fits` in a `data/` directory.
* I would really really like to see some unit tests. I see that @Edenhofer has already [opened an issue](https://github.com/EHTJulia/ScatteringOptics.jl/issues/23) about this. This would make is easier to verify the functionality of the software, and have confidence that everything is working as expected.
* Some community guidelines should be included for third parties who want to contribute or raise issues
