+++ paper/paper.md
@@ -54,28 +54,28 @@ This has driven the development of a theoretical framework [@Johnson_Narayan_201
 The package leverages the strengths of the Julia programming language, which is designed for high-performance computing. A major advantage of Julia is its speed. Programs written in Julia are compiled into efficient native machine code, offering speed comparable to optimized C/C++ or Fortran, and often achieving performance more than a hundred times faster than Python. This speed provides a factor of 10-100 acceleration compared to the current standard implementation of scattering models in Python. This acceleration significantly enhances current state-of-the-art techniques for deriving the maximum posteriori estimate of joint models, including the sky model and phase screens [@Johnson_2016], and also enables the utilization of novel Bayesian inference techniques (available in `Comrade.jl` and related libraries) for deriving full posterior distributions of scattering parameters.
 
 # Mathematics
-`ScatteringOptics.jl` implements a single thin-screen scattering model described in @Johnson_2018 and @Psaltis_2018 that simulates both diffractive and refractive scattering. In many instances, the properties of the interstellar scattering can be well described by a single, thin phase-changing screen $\phi_r(r)$, where $r$ is a transverse coordinate on the screen. The statistical characteristics of scattering can be described by those of the phase screen through its spatial structure function $D_\phi(r)$.
+`ScatteringOptics.jl` implements a single thin-screen scattering model described in @Johnson_2018 and @Psaltis_2018 that simulates both diffractive and refractive scattering. In many instances, the properties of the interstellar scattering can be well described by a single, thin phase-changing screen $\phi(r)$, where $r$ is a two-dimensional transverse coordinate on the screen. The statistical characteristics of scattering can be described by those of the phase screen through its spatial structure function $D_\phi(r)$.
 
 Diffractive scattering causes the angular broadening of the source image. 
 The diffractively scattered image ${I_{ea}}(r)$ is mathematically given by the convolution of the source image ${I_{src}}(r)$ with a blurring scattering kernel, ${G}(r)$, 
 
 $${I_{ea}}(r) = {I_{src}}(r) * {G}(r),$$
 
-where $r$ refers to the two-dimensional phase screen coordinate vector. In radio interferometry, each set of measurements, so-called visibilities, obtained with a pair of antennas at different time and frequency segments, samples a Fourier component of the sky image. The source visibilities, $V_{src}(b)$, are related to the diffractively scattered visibilities, $V_{ea}(b)$, by,
+In radio interferometry, each set of measurements, so-called visibilities, obtained with a pair of antennas at different time and frequency segments, samples a Fourier component of the sky image. The source visibilities, $V_{src}(b)$, are related to the diffractively scattered visibilities, $V_{ea}(b)$, by,
 
 $$V_{ea}(b) = V_{src}(b)\,\text{exp}\left[-\frac{1}{2} D_\phi\left(\frac{b}{1+M}\right)\right],$$
 
-in which $b$ is the baseline vector between observing stations. The magnification $M=D/R$ is the ratio of earth-screen distance $D$ to screen-source distance $R$. 
+in which $b$ is the baseline vector between observing stations. The magnification $M=D/R$ is the ratio of earth-scattering distance $D$ (i.e., the distance between earth and the scattering material) to scattering-source distance $R$. 
 
 Refractive scattering further introduces compact substructures on the diffractively-scattered, angular-broadened images. 
[…]
