### 2024-04-25 02f299a
CompatHelper: bump compat for ComradeBase to 0.7, (keep existing compat)
