### 2024-09-25 780620c
Update math.md
