+++ docs/src/math.md
@@ -4,28 +4,37 @@ CurrentModule = ScatteringOptics
 # Background
 
 ## Mathematics
-`ScatteringOptics.jl` implements a single thin-screen scattering model (described in  [1] and [2]) that simulates both diffractive and refractive scattering. In many instances, the properties of the interstellar scattering can be well described by a single, thin phase-changing screen $\phi_r(r)$, where $r$ is a transverse coordinate on the screen. The statistical characteristics of scattering can be described by those of the phase screen through its spatial structure function $D_\phi(r)$.
+`ScatteringOptics.jl` implements a single thin-screen scattering model (described in  [1] and [2]) that simulates both diffractive and refractive scattering. In many instances, the properties of the interstellar scattering can be well described by a single, thin phase-changing screen $\phi(\vec{r})$, where $\vec{r}$ is a transverse coordinate on the screen. 
+
+#### Spatial Structure Function
+The statistical characteristics of scattering can be described by those of the phase screen through its spatial structure function $D_\phi(\vec{r})$. This function measures the second-order changes in the phase, $\phi(\vec{r})$ of two radio waves separated by a transverse distance $\vec{r}$ when they pass through the screen [2]:
+
+$$D_\phi(r) = \langle [\phi(\vec{r}_0+\vec{r})-\phi(\vec{r}_0)]^2 \rangle$$
+
+#### Diffractive Scattering Kernel
 
 Diffractive scattering causes the angular broadening of the source image. 
-The diffractively scattered image ${I_{ea}}(r)$ is mathematically given by the convolution of the source image ${I_{src}}(r)$ with a blurring scattering kernel, ${G}(r)$, 
+The diffractively scattered image ${I_{ea}}(\vec{r})$ is mathematically given by the convolution of the source image ${I_{src}}(r)$ with a blurring scattering kernel, ${G}(\vec{r})$, 
 
-$${I_{ea}}(r) = {I_{src}}(r) * {G}(r),$$
+$${I_{ea}}(\vec{r}) = {I_{src}}(\vec{r}) * {G}(\vec{r}),$$
 
-where $r$ refers to the two-dimensional phase screen coordinate vector. In radio interferometry, each set of measurements, so-called visibilities, obtained with a pair of antennas at different time and frequency segments, samples a Fourier component of the sky image. The source visibilities, $V_{src}(b)$, are related to the diffractively scattered visibilities, $V_{ea}(b)$, by,
+where $r$ refers to the two-dimensional phase screen coordinate vector. While the above equation is defined in image space, ScatteringOptics.jl performs scattering in Fourier space, where the kernel can be described analytically. In radio interferometry, each set of measurements, so-called visibilities, obtained with a pair of antennas at different time and frequency segments, samples a Fourier component of the sky image. The source visibilities, $V_{src}(\vec{b})$, are related to the diffractively scattered visibilities, $V_{ea}(\vec{b})$, by
 
-$$V_{ea}(b) = V_{src}(b)\,\text{exp}\left[-\frac{1}{2} D_\phi\left(\frac{b}{1+M}\right)\right],$$
+$$V_{ea}(\vec{b}) = V_{src}(\vec{b})\*\text{exp}\left[-\frac{1}{2} D_\phi\left(\frac{\vec{b}}{1+M}\right)\right],$$
 
-in which $b$ is the baseline vector between observing stations. The magnification $M=D/R$ is the ratio of earth-screen distance $D$ to screen-source distance $R$. 
[…]
