+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"math/","page":"Background","title":"Background","text":"CurrentModule = ScatteringOptics","category":"page"},{"location":"math/#Background","page":"Background","title":"Background","text":"","category":"section"},{"location":"math/#Mathematics","page":"Background","title":"Mathematics","text":"","category":"section"},{"location":"math/","page":"Background","title":"Background","text":"ScatteringOptics.jl implements a single thin-screen scattering model (described in  [1] and [2]) that simulates both diffractive and refractive scattering. In many instances, the properties of the interstellar scattering can be well described by a single, thin phase-changing screen phi_r(r), where r is a transverse coordinate on the screen. The statistical characteristics of scattering can be described by those of the phase screen through its spatial structure function D_phi(r).","category":"page"},{"location":"math/","page":"Background","title":"Background","text":"Diffractive scattering causes the angular broadening of the source image.  The diffractively scattered image I_ea(r) is mathematically given by the convolution of the source image I_src(r) with a blurring scattering kernel, G(r), ","category":"page"},{"location":"math/","page":"Background","title":"Background","text":"I_ea(r) = I_src(r) * G(r)","category":"page"},{"location":"math/","page":"Background","title":"Background","text":"where r refers to the two-dimensional phase screen coordinate vector. In radio interferometry, each set of measurements, so-called visibilities, obtained with a pair of antennas at different time and frequency segments, samples a Fourier component of the sky image. The source visibilities, V_src(b), are related to the diffractively scattered visibilities, V_ea(b), by,","category":"page"},{"location":"math/","page":"Background","title":"Background","text":"V_ea(b) = V_src(b)textexpleft-frac12 D_phileft(fracb1+Mright)right","category":"page"},{"location":"math/","page":"Background","title":"Background","text":"in which b is the baseline vector between observing stations. The magnification M=DR is the ratio of earth-screen distance D to screen-source distance R. ","category":"page"},{"location":"math/","page":"Background","title":"Background","text":"Refractive scattering further introduces compact substructures on the diffractively-scattered, angular-broadened images.  The compact substructures arise from phase gradients on the scattering screen nabla phi_r(r). The refractively scattered image I_a(r) is given by","category":"page"},{"location":"math/","page":"Background","title":"Background","text":"I_a(r) approx I_ea(r + r_F^2 nabla phi_r(r))","category":"page"},{"location":"math/","page":"Background","title":"Background","text":"in which the Fresnel scale, r_F = sqrtfracDRD+Rfraclambda2pi is dependent on the observing wavelength lambda [@JohnsonNarayan2016]. ","category":"page"},{"location":"math/","page":"Background","title":"Background","text":"ScatteringOptics.jl implements three analytic probabilistic models for the phase screen phi_r(r), named Dipole, Periodic Boxcar, and Von Mises models in @Psaltis2018, providing the corresponding semi-analytic descriptions of the phase structure function D\\phi(r)$. The default model is the Dipole model, known to be consistent with multi-frequency measurements of Sgr A* [1].","category":"page"},{"location":"math/#References","page":"Background","title":"References","text":"","category":"section"},{"location":"math/
[…]
