### 2024-09-26 58d7de0
Update paper.md
