### 2024-10-07 #6354 @kazuakiyama
Dear Dan,

Thanks for the contact. We were about to reach out to you because @annatartaglia was now fully coming back to work. We have been working on the revision, and will fully revise the documentation and the entire package within few weeks. Do you think it is possible to set the deadline to 4 weeks from now, Monday, November 4th?

Again thank you very much for your patience. Please let us know if this doesn’t work or you have any questions about the paper.

Thanks, Kazu

> On Oct 5, 2024, at 10:28 AM, Dan Foreman-Mackey ***@***.***> wrote:
>
>
> @kazuakiyama <https://github.com/kazuakiyama>, @annatartaglia <https://github.com/annatartaglia> — Please let me know when you'll be able to get back to this review. If you don't think you'll be able to get to it in the next few weeks, we're going to need to reject the submission because we can't just sit on it for such a long time. Please let me know how you would like to proceed and if I don't hear back from you in the next week I will reject the submission. Thanks!
>
