### 2024-10-12 b9a05b1
Update the documentation. 1) switch Documenter.jl to DocumenterVitepress.jl 2) simplify the intro page 3) the background page
