+++ docs/make.jl
@@ -1,22 +1,28 @@
 using ScatteringOptics
 using Documenter
+using DocumenterVitepress
 
-DocMeta.setdocmeta!(ScatteringOptics, :DocTestSetup, :(using ScatteringOptics); recursive=true)
+#DocMeta.setdocmeta!(ScatteringOptics, :DocTestSetup, :(using ScatteringOptics); recursive=true)
 
 makedocs(;
     modules=[ScatteringOptics],
     authors="Anna Tartaglia, Kazunori Akiyama",
     repo="https://github.com/EHTJulia/ScatteringOptics.jl/blob/{commit}{path}#{line}",
     sitename="ScatteringOptics.jl",
-    format=Documenter.HTML(;
-        prettyurls=get(ENV, "CI", "false") == "true",
-        canonical="https://EHTJulia.github.io/ScatteringOptics.jl",
-        edit_link="main",
-        assets=String[],
+#    format=Documenter.HTML(;
+#        prettyurls=get(ENV, "CI", "false") == "true",
+#        canonical="https://EHTJulia.github.io/ScatteringOptics.jl",
+#        edit_link="main",
+#        assets=String[],
+#    ),
+    format = MarkdownVitepress(
+        repo = "https://github.com/EHTJulia/ScatteringOptics.jl",
+        devurl = "dev",
+        deploy_url = "EHTJulia.github.io/ScatteringOptics.jl",
     ),
     pages=[
         "Home" => "index.md",
-        "Background" => "math.md",
+        "Introduction" => "math.md",
         "Tutorial" => "tutorial.md",
         "ScateringOptics.jl API" => "autodocs.md",
     ],
@@ -25,4 +31,5 @@ makedocs(;
 deploydocs(;
     repo="github.com/EHTJulia/ScatteringOptics.jl",
     devbranch="main",
+    push_preview=true,
 )
