+++ docs/src/math.md
@@ -1,48 +1,65 @@
 ```@meta
 CurrentModule = ScatteringOptics
 ```
-# Background
+# Brief Introduction to interstellar scattering
+`ScatteringOptics.jl` implements a single thin-screen intersteller scattering model based on a fast but accurate semi-analytic framework developed in Johnson \& Narayan 2016 [1] and further extended in later work [2-4]. 
 
-## Mathematics
-`ScatteringOptics.jl` implements a single thin-screen scattering model (described in  [1] and [2]) that simulates both diffractive and refractive scattering. In many instances, the properties of the interstellar scattering can be well described by a single, thin phase-changing screen $\phi(\vec{r})$, where $\vec{r}$ is a transverse coordinate on the screen. 
+## What is the interstellar scattering? Why is it important?
+Scintillation is a well-known phenomenon in astronomy. In radio wavelengths, electron density fluctuations in the ionized interstellar plasma cause scattering of radio waves, resulting in the scintillation of compact radio sources in the sky [5, 6]. The interstellar scintillation produces temporal and spectral modulations (i.e., twinkling) in the brightness of objects, as well as distortion of the source images. 
 
-#### Spatial Structure Function
-The statistical characteristics of scattering can be described by those of the phase screen through its spatial structure function $D_\phi(\vec{r})$. This function measures the second-order changes in the phase, $\phi(\vec{r})$ of two radio waves separated by a transverse distance $\vec{r}$ when they pass through the screen [2]:
+Scattering has become increasingly important in high-angular-resolution studies of compact radio sources at micro-arcsecond scales using very long baseline interferometry (VLBI). 
+Notable examples include black hole imaging of the Milky Way's supermassive black hole, Sagittarius A* (Sgr A*), with the Event Horizon Telescope, and studies of extremely high brightness temperature emissions in active galactic nuclei (AGNs) using space VLBI, exemplified by projects like RadioAstron. 
+In both scenarios, the observed interferometric data is influenced by both diffractive and refractive scattering effects. 
+
+## Scattering regimes in the scope
+The model implemented in `ScatteringOptics.jl` will simulate both *diffractive* and *refractive* scattering in the *average* and *ensemble-average* regimes. Here let us briefly introduce each of scattering regimes and also their averaging regimes. 
+
+The interstellar scattering seen in radio wavelengths is in the *strong scattering limit*, where scattering effects are dominated by phase fluctuations on two widely separated scales: **diffractvive** and **refractive** [5, 6]. 
+- **Diffractive scattering** arises from small-scale fluctuations, typically on scales smaller than microarcseconds. Consequently, diffractive effects are quenched for the vast majority of known sources on the sky, and they are typically only seen in extremely compact objects, such as pulsars and fast radio bursts (FRBs). The diffractive scintillation has a coherence timescale of typically only seconds to minutes, and often well time-averaged during the time scales of a majority of astronomical measurements (except for observations of pulsars and FRBs). As a result, typically the effects of the diffractive scattering appear as a angular broadening of the source image.
[…]
