### 2024-10-12 4e0edd5
update documentations to correct typos in math.md. Also add top and introduction pages.
