+++ docs/src/math.md
@@ -1,46 +1,52 @@
 ```@meta
 CurrentModule = ScatteringOptics
 ```
-# Brief Introduction to interstellar scattering
-`ScatteringOptics.jl` implements a single thin-screen intersteller scattering model based on a fast but accurate semi-analytic framework developed in Johnson \& Narayan 2016 [1] and further extended in later work [2-4]. 
+## Brief Introduction to Interstellar Scattering
 
-## What is the interstellar scattering? Why is it important?
-Scintillation is a well-known phenomenon in astronomy. In radio wavelengths, electron density fluctuations in the ionized interstellar plasma cause scattering of radio waves, resulting in the scintillation of compact radio sources in the sky [5, 6]. The interstellar scintillation produces temporal and spectral modulations (i.e., twinkling) in the brightness of objects, as well as distortion of the source images. 
+`ScatteringOptics.jl` implements a single thin-screen interstellar scattering model based on a fast yet accurate semi-analytic framework developed by Johnson & Narayan (2016) [1], and further extended in subsequent works [2-4].
 
-Scattering has become increasingly important in high-angular-resolution studies of compact radio sources at micro-arcsecond scales using very long baseline interferometry (VLBI). 
-Notable examples include black hole imaging of the Milky Way's supermassive black hole, Sagittarius A* (Sgr A*), with the Event Horizon Telescope, and studies of extremely high brightness temperature emissions in active galactic nuclei (AGNs) using space VLBI, exemplified by projects like RadioAstron. 
-In both scenarios, the observed interferometric data is influenced by both diffractive and refractive scattering effects. 
 
-## Scattering regimes in the scope
-The model implemented in `ScatteringOptics.jl` will simulate both *diffractive* and *refractive* scattering in the *average* and *ensemble-average* regimes. Here let us briefly introduce each of scattering regimes and also their averaging regimes. 
+## Scattering Regimes in Scope
 
-The interstellar scattering seen in radio wavelengths is in the *strong scattering limit*, where scattering effects are dominated by phase fluctuations on two widely separated scales: **diffractvive** and **refractive** [5, 6]. 
-- **Diffractive scattering** arises from small-scale fluctuations, typically on scales smaller than microarcseconds. Consequently, diffractive effects are quenched for the vast majority of known sources on the sky, and they are typically only seen in extremely compact objects, such as pulsars and fast radio bursts (FRBs). The diffractive scintillation has a coherence timescale of typically only seconds to minutes, and often well time-averaged during the time scales of a majority of astronomical measurements (except for observations of pulsars and FRBs). As a result, typically the effects of the diffractive scattering appear as a angular broadening of the source image.
-- **Refractive scattering** arises from large-scale fluctuations, typically on scales of micro to milli-arcseconds depending on the location of the sky and observing frequency. typically days to months. Refractive effects are only quenched for a source that has an angular size larger than those scales. Consequently, they are present in many compact objects such as active galactic nuclei (AGNs). Refractive scintillation is more persistent and has a coherence timescale of days to months. Refractive scattering introduces *refractive substructures* into the observed angular-broadened image.
+The model implemented in `ScatteringOptics.jl` simulates both *diffractive* and *refractive* scattering in the *average* and *ensemble-average* regimes. Below is a brief introduction to each scattering regime and their corresponding averaging regimes.
 
-The late 80's work by Ramesh Narayan and Jeremy J. Goodman [7, 8] showed that there are three distinct averaging regimes for images in the strong-scattering limit.
[…]
