### 2024-10-12 101d5c2
update the compatibility info for Comrade and ComradeBase
