+++ Project.toml
@@ -18,7 +18,7 @@ StationaryRandomFields = "6ab34832-f7e5-40a4-9cd7-47ea82b5c144"
 VLBISkyModels = "d6343c73-7174-4e0f-bb64-562643efbeca"
 
 [compat]
-ComradeBase = "0.5, 0.6, 0.7, 0.8"
+ComradeBase = "0.8"
 DocStringExtensions = "0.9"
 EHTUtils = "0.1"
 FFTW = "1"
@@ -28,8 +28,8 @@ NonlinearSolve = "1, 3"
 QuadGK = "2"
 SpecialFunctions = "2"
 StationaryRandomFields = "0.1"
-VLBISkyModels = "0.2, 0.3, 0.4, 0.5, 0.6"
-julia = "1.9"
+VLBISkyModels = "0.6"
+julia = "1.11"
 
 [extras]
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
