+++ dev/assets/introduction.md.zm5VhivB.lean.js
@@ -1,0 +1,2 @@
+import{_ as t,c as i,a5 as a,o as n}from"./chunks/framework.CllLk_H8.js";const g=JSON.parse('{"title":"Introduction","description":"","frontmatter":{},"headers":[],"relativePath":"introduction.md","filePath":"introduction.md","lastUpdated":null}'),s={name:"introduction.md"};function r(o,e,l,c,h,d){return n(),i("div",null,e[0]||(e[0]=[a(`<h1 id="introduction" tabindex="-1">Introduction <a class="header-anchor" href="#introduction" aria-label="Permalink to &quot;Introduction&quot;">​</a></h1><p><a href="https://github.com/EHTJulia/ScatteringOptics.jl" target="_blank" rel="noreferrer"><code>ScatteringOptics.jl</code></a> is an astronomy software package developed in the Julia programming language for interstellar scintillation observed at radio wavelengths. It implements physical models for the anisotropic scattering of radio waves, which arise from turbulence in the ionized interstellar medium.</p><p>This toolkit is aiming to provide capabilities of simulating and modeling the temporal, spatial, and spectral effects of interstellar scintillation primarily in the strong scattering regime, taking advantage of Julia&#39;s speed and composability. The package is designed to work with sky models and interferometric data types from the advanced Bayesian radio interferometric modeling package <a href="https://github.com/ptiede/Comrade.jl" target="_blank" rel="noreferrer"><code>Comrade.jl</code></a>.</p><p>The package provides essential functionalities for modeling, analyzing, and interpreting the images of the Galactic Center&#39;s supermassive black hole, Sagittarius A*, especially with the Event Horizon Telescope, and also the images of extremely high brightness temperature emissions in active galactic nuclei using space very long baseline interferometry.</p><h2 id="installation" tabindex="-1">Installation <a class="header-anchor" href="#installation" aria-label="Permalink to &quot;Installation&quot;">​</a></h2><p>The package is registered in the Julia standard repository, and installable through the standard ways by</p><div class="language-julia vp-adaptive-theme"><button title="Copy Code" class="copy"></button><span class="lang">julia</span><pre class="shiki shiki-themes github-light github-dark vp-code" tabindex="0"><code><span class="line"><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">using</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;"> Pkg</span></span>
[…]
