### 2024-10-12 f465d13
update the workflow script for JOSS paper
