### 2024-10-12 364639c
fix a typo in draft-pdf.yml
