+++ .github/workflows/draft-pdf.yml
@@ -1,4 +1,5 @@
-on: [push]
+name: Draft JOSS Paper PDF
+on:
   push:
     paths:
       - paper/**
@@ -16,7 +17,7 @@ jobs:
         with:
           journal: joss
           # This should be the path to the paper within your repo.
-          paper-path: paper/paper.md
+          paper-path: paper.md
       - name: Upload
         uses: actions/upload-artifact@v4
         with:
@@ -24,4 +25,4 @@ jobs:
           # This is the output path where Pandoc will write the compiled
           # PDF. Note, this should be the same directory as the input
           # paper.md
-          path: paper/paper.pdf
+          path: paper.pdf
