+++ dev/assets/autodocs.md.D4PlnnB4.js
@@ -1,0 +1,1 @@
+import{_ as r,c as l,a5 as s,j as i,a,G as n,B as p,o as c}from"./chunks/framework.CllLk_H8.js";const W=JSON.parse('{"title":"ScateringOptics.jl API","description":"","frontmatter":{},"headers":[],"relativePath":"autodocs.md","filePath":"autodocs.md","lastUpdated":null}'),o={name:"autodocs.md"},d={class:"jldocstring custom-block",open:""},h={class:"jldocstring custom-block",open:""},g={class:"jldocstring custom-block",open:""},k={class:"jldocstring custom-block",open:""},u={class:"jldocstring custom-block",open:""},b={class:"jldocstring custom-block",open:""},m={class:"jldocstring custom-block",open:""},f={class:"jldocstring custom-block",open:""},y={class:"jldocstring custom-block",open:""},E={class:"jldocstring custom-block",open:""},S={class:"jldocstring custom-block",open:""},j={class:"jldocstring custom-block",open:""},v={class:"jldocstring custom-block",open:""},F={class:"jldocstring custom-block",open:""},O={class:"jldocstring custom-block",open:""},C={class:"jldocstring custom-block",open:""},A={class:"jldocstring custom-block",open:""},x={class:"jldocstring custom-block",open:""},T={class:"jldocstring custom-block",open:""},N={class:"jldocstring custom-block",open:""},M={class:"jldocstring custom-block",open:""},D={class:"jldocstring custom-block",open:""},B={class:"jldocstring custom-block",open:""},_={class:"jldocstring custom-block",open:""},z={class:"jldocstring custom-block",open:""},P={class:"jldocstring custom-block",open:""},w={class:"jldocstring custom-block",open:""},K={class:"jldocstring custom-block",open:""};function L(J,e,H,q,I,R){const t=p("Badge");return c(),l("div",null,[e[85]||(e[85]=s('<h1 id="ScateringOptics.jl-API" tabindex="-1">ScateringOptics.jl API <a class="header-anchor" href="#ScateringOptics.jl-API" aria-label="Permalink to &quot;ScateringOptics.jl API {#ScateringOptics.jl-API}&quot;">​</a></h1><h2 id="index" tabindex="-1">Index <a class="header-anchor" href="#index" aria-label="Permalink to &quot;Index&quot;">​</a></h2><ul><li><a href="#ScatteringOptics.Params_Johnson2018"><code>ScatteringOptics.Params_Johnson2018</code></a></li><li><a href="#ScatteringOptics.AbstractKzetaFinder"><code>ScatteringOptics.AbstractKzetaFinder</code></a></li><li><a href="#ScatteringOptics.AbstractPhaseScreen"><code>ScatteringOptics.AbstractPhaseScreen</code></a></li><li><a href="#ScatteringOptics.AbstractScatteringKernel"><code>ScatteringOptics.AbstractScatteringKernel</code></a></li><li><a href="#ScatteringOptics.AbstractScatteringModel"><code>ScatteringOptics.AbstractScatteringModel</code></a></li><li><a href="#ScatteringOptics.ApproximatedScatteringKernel"><code>ScatteringOptics.ApproximatedScatteringKernel</code></a></li><li><a href="#ScatteringOptics.DipoleScatteringModel"><code>ScatteringOptics.DipoleScatteringModel</code></a></li><li><a href="#ScatteringOptics.Dipole_KzetaFinder"><code>ScatteringOptics.Dipole_KzetaFinder</code></a></li><li><a href="#ScatteringOptics.ExactScatteringKernel"><code>ScatteringOptics.ExactScatteringKernel</code></a></li><li><a h
[…]
