### 2024-10-12 d3b32f6
update visibility_point functions to reflect changes in Comrade v0.8
