+++ src/scatteringmodels/kernels/approx.jl
@@ -25,13 +25,6 @@ function radialextent(skm::ApproximatedScatteringKernel{T,S,N}) where {T,S,N}
     return convert(T, 5) * calc_θrad(skm.sm.θmaj) * (ν2λcm(νref) / skm.sm.λ0)^2
 end
 
-# for Comrade before v0.6
-#@inline function visibility_point(skm::ApproximatedScatteringKernel{T,S,N}, u, v, time, freq) where {T,S,N}
-#    ν = iszero(freq) ? skm.νref : freq
-#    return visibility_point_approx(skm.sm, ν2λcm(ν), u, v) + zero(T)im
-#end
-
-# this is for Comrade after v0.8
 @inline function visibility_point(skm::ApproximatedScatteringKernel{T,S,N}, p) where {T,S,N}
     (;U, V, Ti, Fr) = p
     ν = iszero(Fr) ? skm.νref : Fr
