### 2024-10-12 afc888f
reorganize introduction in the documenatation
