+++ dev/assets/chunks/@localSearchIndexroot.BAo3iFJd.js
@@ -1,1 +1,0 @@
-const e='{"documentCount":11,"nextId":11,"documentIds":{"0":"/ScatteringOptics.jl/dev/autodocs#ScateringOptics.jl-API","1":"/ScatteringOptics.jl/dev/autodocs#index","2":"/ScatteringOptics.jl/dev/autodocs#docstrings","3":"/ScatteringOptics.jl/dev/introduction#introduction","4":"/ScatteringOptics.jl/dev/introduction#installation","5":"/ScatteringOptics.jl/dev/introduction#What-are-interstellar-scattering-and-scintillation?-Why-are-they-important?","6":"/ScatteringOptics.jl/dev/math#Brief-Introduction-to-Interstellar-Scattering","7":"/ScatteringOptics.jl/dev/math#Scattering-Regimes-in-Scope","8":"/ScatteringOptics.jl/dev/math#Scattering-Model","9":"/ScatteringOptics.jl/dev/math#references","10":"/ScatteringOptics.jl/dev/tutorial#tutorial"},"fieldIds":{"title":0,"titles":1,"text":2},"fieldLength":{"0":[3,1,1],"1":[1,3,40],"2":[1,3,410],"3":[1,1,100],"4":[1,1,21],"5":[10,1,128],"6":[5,1,34],"7":[4,1,168],"8":[2,1,207],"9":[1,1,84],"10":[1,1,131]},"averageFieldLength":[2.727272727272727,1.3636363636363635,120.36363636363636],"storedFields":{"0":{"title":"ScateringOptics.jl API","titles":[]},"1":{"title":"Index","titles":["ScateringOptics.jl API"]},"2":{"title":"Docstrings","titles":["ScateringOptics.jl API"]},"3":{"title":"Introduction","titles":[]},"4":{"title":"Installation","titles":["Introduction"]},"5":{"title":"What are interstellar scattering and scintillation? Why are they important?","titles":["Introduction"]},"6":{"title":"Brief Introduction to Interstellar Scattering","titles":[]},"7":{"title":"Scattering Regimes in Scope","titles":[]},"8":{"title":"Scattering Model","titles":[]},"9":{"title":"References","titles":[]},"10":{"title":"Tutorial","titles":[]}},"dirtCount":0,"index":[["μas",{"2":{"10":3}}],["995",{"2":{"9":2}}],["963",{"2":{"9":2}}],["≈iea",{"2":{"8":1}}],["∇ϕ",{"2":{"8":1}}],["−12dϕ",{"2":{"8":1}}],["−ϕ",{"2":{"8":1}}],["∗g",{"2":{"8":1}}],["865",{"2":{"9":1}}],["833",{"2":{"9":2}}],["826",{"2":{"9":2}}],["8",{"2":{"7":1,"9":1}}],["74",{"2":{"9":2}}],["7",{"2":{"7":1,"9":1}}],["637x",{"2":{"9":1}}],["6",{"2":{"7":1,"8":1,"9":1}}],["quot",{"2":{"8":2}}],["quenched",{"2":{"7":3}}],["q",{"2":{"2":1}}],["qbar",{"2":{"2":1}}],["hi",{"2":{"10":2}}],["hig
[…]
