+++ dev/assets/chunks/@localSearchIndexroot.BlUtmf4W.js
@@ -1,0 +1,1 @@
+const e='{"documentCount":11,"nextId":11,"documentIds":{"0":"/ScatteringOptics.jl/dev/autodocs#ScateringOptics.jl-API","1":"/ScatteringOptics.jl/dev/autodocs#index","2":"/ScatteringOptics.jl/dev/autodocs#docstrings","3":"/ScatteringOptics.jl/dev/introduction#introduction","4":"/ScatteringOptics.jl/dev/introduction#installation","5":"/ScatteringOptics.jl/dev/introduction#What-are-interstellar-scattering-and-scintillation?-Why-are-they-important?","6":"/ScatteringOptics.jl/dev/tutorial#tutorial","7":"/ScatteringOptics.jl/dev/math#Brief-Introduction-to-Interstellar-Scattering","8":"/ScatteringOptics.jl/dev/math#Scattering-Regimes-in-Scope","9":"/ScatteringOptics.jl/dev/math#Scattering-Model","10":"/ScatteringOptics.jl/dev/math#references"},"fieldIds":{"title":0,"titles":1,"text":2},"fieldLength":{"0":[3,1,1],"1":[1,3,40],"2":[1,3,410],"3":[1,1,100],"4":[1,1,21],"5":[10,1,128],"6":[1,1,131],"7":[5,1,34],"8":[4,1,168],"9":[2,1,207],"10":[1,1,84]},"averageFieldLength":[2.727272727272727,1.3636363636363635,120.36363636363636],"storedFields":{"0":{"title":"ScateringOptics.jl API","titles":[]},"1":{"title":"Index","titles":["ScateringOptics.jl API"]},"2":{"title":"Docstrings","titles":["ScateringOptics.jl API"]},"3":{"title":"Introduction","titles":[]},"4":{"title":"Installation","titles":["Introduction"]},"5":{"title":"What are interstellar scattering and scintillation? Why are they important?","titles":["Introduction"]},"6":{"title":"Tutorial","titles":[]},"7":{"title":"Brief Introduction to Interstellar Scattering","titles":[]},"8":{"title":"Scattering Regimes in Scope","titles":[]},"9":{"title":"Scattering Model","titles":[]},"10":{"title":"References","titles":[]}},"dirtCount":0,"index":[["995",{"2":{"10":2}}],["963",{"2":{"10":2}}],["≈iea",{"2":{"9":1}}],["∇ϕ",{"2":{"9":1}}],["−12dϕ",{"2":{"9":1}}],["−ϕ",{"2":{"9":1}}],["∗g",{"2":{"9":1}}],["865",{"2":{"10":1}}],["833",{"2":{"10":2}}],["826",{"2":{"10":2}}],["8",{"2":{"8":1,"10":1}}],["74",{"2":{"10":2}}],["7",{"2":{"8":1,"10":1}}],["637x",{"2":{"10":1}}],["6",{"2":{"8":1,"9":1,"10":1}}],["μas",{"2":{"6":3}}],["quot",{"2":{"9":2}}],["quenched",{"2":{"8":3}}],["q",{"2":{"2":1}}],["qbar",{"2":{"2":1}}],["hi",{"2":{"6"
[…]
