+++ dev/assets/chunks/@localSearchIndexroot.BU5DmlQl.js
@@ -1,0 +1,1 @@
+const e='{"documentCount":11,"nextId":11,"documentIds":{"0":"/ScatteringOptics.jl/dev/autodocs#ScateringOptics.jl-API","1":"/ScatteringOptics.jl/dev/autodocs#index","2":"/ScatteringOptics.jl/dev/autodocs#docstrings","3":"/ScatteringOptics.jl/dev/math#Brief-Introduction-to-Interstellar-Scattering","4":"/ScatteringOptics.jl/dev/math#Scattering-Regimes-in-Scope","5":"/ScatteringOptics.jl/dev/math#Scattering-Model","6":"/ScatteringOptics.jl/dev/math#references","7":"/ScatteringOptics.jl/dev/introduction#introduction","8":"/ScatteringOptics.jl/dev/introduction#What-are-interstellar-scattering-and-scintillation?-Why-are-they-important?","9":"/ScatteringOptics.jl/dev/introduction#installation","10":"/ScatteringOptics.jl/dev/tutorial#tutorial"},"fieldIds":{"title":0,"titles":1,"text":2},"fieldLength":{"0":[3,1,1],"1":[1,3,40],"2":[1,3,410],"3":[5,1,34],"4":[4,1,168],"5":[2,1,207],"6":[1,1,84],"7":[1,1,100],"8":[10,1,128],"9":[1,1,21],"10":[1,1,131]},"averageFieldLength":[2.727272727272727,1.3636363636363635,120.36363636363636],"storedFields":{"0":{"title":"ScateringOptics.jl API","titles":[]},"1":{"title":"Index","titles":["ScateringOptics.jl API"]},"2":{"title":"Docstrings","titles":["ScateringOptics.jl API"]},"3":{"title":"Brief Introduction to Interstellar Scattering","titles":[]},"4":{"title":"Scattering Regimes in Scope","titles":[]},"5":{"title":"Scattering Model","titles":[]},"6":{"title":"References","titles":[]},"7":{"title":"Introduction","titles":[]},"8":{"title":"What are interstellar scattering and scintillation? Why are they important?","titles":["Introduction"]},"9":{"title":"Installation","titles":["Introduction"]},"10":{"title":"Tutorial","titles":[]}},"dirtCount":0,"index":[["μas",{"2":{"10":3}}],["995",{"2":{"6":2}}],["963",{"2":{"6":2}}],["≈iea",{"2":{"5":1}}],["∇ϕ",{"2":{"5":1}}],["−12dϕ",{"2":{"5":1}}],["−ϕ",{"2":{"5":1}}],["∗g",{"2":{"5":1}}],["865",{"2":{"6":1}}],["833",{"2":{"6":2}}],["826",{"2":{"6":2}}],["8",{"2":{"4":1,"6":1}}],["74",{"2":{"6":2}}],["7",{"2":{"4":1,"6":1}}],["637x",{"2":{"6":1}}],["6",{"2":{"4":1,"5":1,"6":1}}],["quot",{"2":{"5":2}}],["quenched",{"2":{"4":3}}],["q",{"2":{"2":1}}],["qbar",{"2":{"2":1}}],["hi",{"2":{"10":2}}],["hig
[…]

[…]
