### 2024-10-13 #6354 @dfm
Thanks for the update - sounds good! But, like I've mentioned several times now, please don't plan on just revising in one go. It's _much_ more useful to _iterate_ on JOSS reviews! Please start responding to GitHub issues and on this thread ASAP rather than waiting for a complete revision. That's not the best use of anyone's time.
