+++ dev/assets/math.md.J8U844d0.lean.js
@@ -1,1 +1,0 @@
-import{_ as l,c as Q,a5 as s,j as t,a,o as T}from"./chunks/framework.CllLk_H8.js";const x3=JSON.parse('{"title":"","description":"","frontmatter":{},"headers":[],"relativePath":"math.md","filePath":"math.md","lastUpdated":null}'),o={name:"math.md"},r={class:"MathJax",jax:"SVG",style:{direction:"ltr",position:"relative"}},n={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-0.566ex"},xmlns:"http://www.w3.org/2000/svg",width:"4.129ex",height:"2.477ex",role:"img",focusable:"false",viewBox:"0 -845 1825 1095","aria-hidden":"true"},m={class:"MathJax",jax:"SVG",style:{direction:"ltr",position:"relative"}},d={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-0.025ex"},xmlns:"http://www.w3.org/2000/svg",width:"1.02ex",height:"1.937ex",role:"img",focusable:"false",viewBox:"0 -845 451 856","aria-hidden":"true"},i={class:"MathJax",jax:"SVG",style:{direction:"ltr",position:"relative"}},h={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-0.667ex"},xmlns:"http://www.w3.org/2000/svg",width:"5.795ex",height:"2.579ex",role:"img",focusable:"false",viewBox:"0 -845 2561.4 1140","aria-hidden":"true"},p={class:"MathJax",jax:"SVG",style:{direction:"ltr",position:"relative"}},g={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-0.566ex"},xmlns:"http://www.w3.org/2000/svg",width:"4.129ex",height:"2.477ex",role:"img",focusable:"false",viewBox:"0 -845 1825 1095","aria-hidden":"true"},c={class:"MathJax",jax:"SVG",style:{direction:"ltr",position:"relative"}},x={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-0.025ex"},xmlns:"http://www.w3.org/2000/svg",width:"1.02ex",height:"1.937ex",role:"img",focusable:"false",viewBox:"0 -845 451 856","aria-hidden":"true"},w={class:"MathJax",jax:"SVG",display:"true",style:{direction:"ltr",display:"block","text-align":"center",margin:"1em 0",position:"relative"}},u={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-0.667ex"},xmlns:"http://www.w3.org/2000/svg",width:"29.603ex",height:"2.667ex",role:"img",focusable:"false",viewBox:"0 -883.9 13084.5 1178.9","aria-hidden":"true"},f={class:"MathJax",jax:"SVG",style:{direction:"ltr",position:"relative"}},H={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-0.566ex"},xmlns:"http://www.w3.org/2000/svg",width:"5.556ex",height:"2.477ex",role:"img",focusable:"false",viewBox:"0 -845 2455.6 1095","aria-hidden":"true"},y={class:"MathJax",jax:"SVG",style:{direction:"ltr",position:"relative"}},k={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-0.566ex"},xmlns:"http://www.w3.org/2000/svg",width:"6.128ex",height:"2.477ex",role:"img",focusable:"false",viewBox:"0 -845 2708.7 1095","aria-hidden":"true"},M={class:"M
[…]
