### 2024-10-16 9947e6e
update the documentation. Some new pages are added based on the JOSS review
