+++ Project.toml
@@ -29,7 +29,7 @@ QuadGK = "2"
 SpecialFunctions = "2"
 StationaryRandomFields = "0.1"
 VLBISkyModels = "0.6"
-julia = "1.11"
+julia = "1.10"
 
 [extras]
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"

[…]
