+++ src/scatteringmodels/kernels/exact.jl
@@ -25,7 +25,8 @@ function radialextent(skm::ExactScatteringKernel{T,S,N}) where {T,S,N}
 end
 
 @inline function visibility_point(skm::ExactScatteringKernel{T,S,N}, p) where {T,S,N}
-    (;U, V, Ti, Fr) = p
-    ν = iszero(Fr) ? skm.νref : Fr
+    U = p.U
+    V = p.V
+    ν = :Fr in keys(p) ? p.Fr : skm.νref
     return visibility_point_exact(skm.sm, ν2λcm(ν), U, V) + zero(T)im
 end
