### 2024-10-16 fe92620
update the data structure of IntensityMap
