### 2024-10-16 c8d7f9f
allow image_vis to specify a specific scattering screen
