+++ docs/src/tutorials/gettingstarted.md
@@ -3,59 +3,102 @@ CurrentModule = ScatteringOptics
 ```
 
 # Getting Started
-This example code segment uses ScatteringOptics.jl to simulate interstellar scattering on an input [Comrade.jl](https://github.com/ptiede/Comrade.jl) `SkyModel` object via the `image_scatter()` function.
+`ScatteringOptics.jl` is designed as a package inside the ecosystem of [`Comrade.jl`](https://github.com/ptiede/Comrade.jl). The scattering model in the package can scatter any sky model types from `Comrade.jl`.
 
-```julia
-using ScatteringOptics
-using EHTImages
-using PythonPlot
-using EHTUtils
+## Loading your image
+Here, we use an example image in [`eht-imaging`](https://github.com/achael/eht-imaging). Data can be downloaded from [here](data/jason_mad_eofn.fits). This is a general relativistic magnetohydrodynamic (GRMHD) model of the magnetic arrestic disk originally from [Dexter et al. 2020](https://ui.adsabs.harvard.edu/abs/2020MNRAS.494.4168D/abstract).
+
+```@example 1
+using CairoMakie
+# This is the base package for the Skymodel of the Comrade.jl ecosystem
+using VLBISkyModels
+# Alternatively, you can import Comrade.jl instead. Either works.
+# using Comrade
 
 # Load a model image FITS file
-im = load_fits("jason_mad_eofn.fits")
-# Initialize a comrade SkyModel from the loaded image
-imap = intensitymap(im)
+im = load_fits("data/jason_mad_eofn.fits", IntensityMap)
+
 # Plot source image
-imshow(im, angunit=EHTUtils.μas)
+imageviz(im, size=(600, 500), colormap=:afmhot)
 ```
-![fig](images/src.png)
 
-Optionally, users can retrieve a purely diffractive scattering ensemble average image with the `convolve` function in `VLBISkyModels.jl`.
+You can instantiate your scattering model with `ScatteringModel()`. If nothing is specified, the model will use the best-fit parameter set for Sgr A* in [Johnson et al. 2018](https://ui.adsabs.harvard.edu/abs/2018ApJ...865..104J/abstract). 
+```@example 1
+using ScatteringOptics
 
-```julia
-# Initialize a scattering model with desired scattering paramaters, otherwise default ISM parameters are used
-# The default model is a Dipole model
+# initialize the scattering model
 sm = ScatteringModel()
+```
+You can change the parameters if you want to simulate a different scattering screen. See [ScatteringOptics.DipoleScatteringModel](@ref) for arguments. 
+
+## Simulate diffractive scattering 
+As explained in [Brief Introduction to Interstellar Scattering](@ref), diffractive scattering will cause angular broaderning of the resultant image, which is described by the convolution of the source image with a scattering kernel. `ScatteringOptics.jl` implements a `Comrade.jl`'s skymodel of the scattering kernel. You can generate it by 
+
+```@example 1
+# Frequency of the image
+νref = metadata(im).frequency
+print("Frequency of the image: ", νref/1e9," GHz")
 
 # Create kernel from scattering model
-skm = kernelmodel(sm, νref=230e9)
-
-# Convolve to get the ensemble averaged image
-using VLBISkyModels
-imap_ea = convolve(imap, skm)
-
-# Map and plot convolved image
-im_ea = copy(im)
-map!(im_ea, imap_ea)
-imshow(im_ea, angunit=EHTUtils.μas)
+skm = kernelmodel(sm, νref=νref)
 ```
-![fig](images/ea.png)
 
[…]
