+++ dev/assets/chunks/@localSearchIndexroot.D24OaoZ6.js
@@ -1,0 +1,1 @@
+const e='{"documentCount":18,"nextId":18,"documentIds":{"0":"/ScatteringOptics.jl/dev/api#ScateringOptics.jl-API","1":"/ScatteringOptics.jl/dev/api#index","2":"/ScatteringOptics.jl/dev/api#docstrings","3":"/ScatteringOptics.jl/dev/introduction#introduction","4":"/ScatteringOptics.jl/dev/introduction#What-are-interstellar-scattering-and-scintillation?-Why-are-they-important?","5":"/ScatteringOptics.jl/dev/introduction#installation","6":"/ScatteringOptics.jl/dev/math#Brief-Introduction-to-Interstellar-Scattering","7":"/ScatteringOptics.jl/dev/math#Scattering-Regimes-in-Scope","8":"/ScatteringOptics.jl/dev/math#Scattering-Model","9":"/ScatteringOptics.jl/dev/math#references","10":"/ScatteringOptics.jl/dev/tutorials/custommodels#Define-Your-Own-Scattering-Model","11":"/ScatteringOptics.jl/dev/tutorials/custommodels#references","12":"/ScatteringOptics.jl/dev/tutorials/gettingstarted#Getting-Started","13":"/ScatteringOptics.jl/dev/tutorials/gettingstarted#Loading-your-image","14":"/ScatteringOptics.jl/dev/tutorials/gettingstarted#Simulate-diffractive-scattering","15":"/ScatteringOptics.jl/dev/tutorials/gettingstarted#Simulating-Refractive-Scattering","16":"/ScatteringOptics.jl/dev/tutorials/nondefaultmodels#Use-Non-default-Models","17":"/ScatteringOptics.jl/dev/tutorials/nondefaultmodels#references"},"fieldIds":{"title":0,"titles":1,"text":2},"fieldLength":{"0":[3,1,1],"1":[1,3,40],"2":[1,3,441],"3":[1,1,100],"4":[10,1,128],"5":[1,1,21],"6":[5,1,34],"7":[4,5,168],"8":[2,5,255],"9":[1,5,84],"10":[5,1,254],"11":[1,5,39],"12":[2,1,23],"13":[3,2,136],"14":[3,2,166],"15":[3,2,97],"16":[4,1,95],"17":[1,4,36]},"averageFieldLength":[2.8333333333333335,2.4444444444444446,117.66666666666667],"storedFields":{"0":{"title":"ScateringOptics.jl API","titles":[]},"1":{"title":"Index","titles":["ScateringOptics.jl API"]},"2":{"title":"Docstrings","titles":["ScateringOptics.jl API"]},"3":{"title":"Introduction","titles":[]},"4":{"title":"What are interstellar scattering and scintillation? Why are they important?","titles":["Introduction"]},"5":{"title":"Installation","titles":["Introduction"]},"6":{"title":"Brief Introduction to Interstellar Scattering","titles":[]},"7":{"title":"Scattering Regimes in Scope","titles":["Brief Introduction 
[…]
