+++ dev/assets/tutorials_gettingstarted.md.ehCANCHY.lean.js
@@ -1,37 +1,0 @@
-import{_ as i,c as a,a5 as n,o as t}from"./chunks/framework.DWm5V20y.js";const e="/ScatteringOptics.jl/dev/assets/src.B9lzDg5m.png",l="/ScatteringOptics.jl/dev/assets/ea.CxqHbo9L.png",h="/ScatteringOptics.jl/dev/assets/avg.DBbbsHhQ.png",o=JSON.parse('{"title":"Getting Started","description":"","frontmatter":{},"headers":[],"relativePath":"tutorials/gettingstarted.md","filePath":"tutorials/gettingstarted.md","lastUpdated":null}'),p={name:"tutorials/gettingstarted.md"};function k(r,s,d,E,g,c){return t(),a("div",null,s[0]||(s[0]=[n(`<h1 id="Getting-Started" tabindex="-1">Getting Started <a class="header-anchor" href="#Getting-Started" aria-label="Permalink to &quot;Getting Started {#Getting-Started}&quot;">​</a></h1><p>This example code segment uses ScatteringOptics.jl to simulate interstellar scattering on an input <a href="https://github.com/ptiede/Comrade.jl" target="_blank" rel="noreferrer">Comrade.jl</a> <code>SkyModel</code> object via the <code>image_scatter()</code> function.</p><div class="language-julia vp-adaptive-theme"><button title="Copy Code" class="copy"></button><span class="lang">julia</span><pre class="shiki shiki-themes github-light github-dark vp-code" tabindex="0"><code><span class="line"><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">using</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;"> ScatteringOptics</span></span>
-<span class="line"><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">using</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;"> EHTImages</span></span>
-<span class="line"><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">using</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;"> PythonPlot</span></span>
-<span class="line"><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">using</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;"> EHTUtils</span></span>
-<span class="line"></span>
-<span class="line"><span style="--shiki-light:#6A737D;--shiki-dark:#6A737D;"># Load a model image FITS file</span></span>
-<span class="line"><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">im </span><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">=</span><span style="--shiki-light:#005CC5;--shiki-dark:#79B8FF;"> load_fits</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">(</span><span style="--shiki-light:#032F62;--shiki-dark:#9ECBFF;">&quot;jason_mad_eofn.fits&quot;</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">)</span></span>
-<span class="line"><span style="--shiki-light:#6A737D;--shiki-dark:#6A737D;"># Initialize a comrade SkyModel from the loaded image</span></span>
[…]

[…]
