+++ v0.1.3/assets/index.md.hJYNFSs7.js
@@ -1,0 +1,1 @@
+import{_ as t,c as e,o as a}from"./chunks/framework.D0qTNtaw.js";const p=JSON.parse('{"title":"","description":"","frontmatter":{"layout":"home","hero":{"name":"ScatteringOptics.jl","tagline":"A Julia package to model and simulate interstellar scattering and scintillation","actions":[{"theme":"brand","text":"Introduction","link":"/introduction"},{"theme":"alt","icon":"github","text":"View on Github","link":"https://github.com/EHTJulia/ScatteringOptics.jl"},{"theme":"alt","text":"Tutorial","link":"/tutorial"}]}},"headers":[],"relativePath":"index.md","filePath":"index.md","lastUpdated":null}'),i={name:"index.md"};function n(o,l,r,c,s,d){return a(),e("div")}const u=t(i,[["render",n]]);export{p as __pageData,u as default};
