### 2024-10-16 a619d55
add a clarification for the typo made in Psaltis et al. 2018
