+++ src/kzetafinders/periodicboxcar.jl
@@ -14,6 +14,8 @@ end
 
 
 # Equation 47 of Psaltis et al. 2018.
+#    Note: Eq 47 has a typo and it has an extra kzeta. We use the correct one following
+#          Psaltis+2018's reference implementation in the eht-imaging library.
 @inline function kzetafinder_equation(kzeta, finder::PeriodicBoxCar_KzetaFinder)
     #x = π ./ (1 .+ kzeta)
     #return sin.(x) ./ x .- finder.ζ0
