### 2024-10-16 EHTJulia/ScatteringOptics.jl#22 @kazuakiyama
We are very sorry that our response came very late. @annatartaglia had a career transition after we received the report and unfortunately, it took a long time for us to come back to work on the revision. The package revision is almost done and here we would like to answer your question.

> The referenced formula (Equation 47 of Psaltis et al. 2018) contains an additional kzeta in the argument to sinus. Is that potentially missing here?
Thanks for carefully going through the equation. Indeed, this implementation is slightly different from Equation 47 of Psaltis et al. 2018. To our knowledge, Equation 47 has a typo --- this extra `kzeta` you found is not required. Unfortunately, for some reason, the paper didn't complete the review process or have another revision, and the typo has been left in the arxiv posting for a while.

You can find the correct model implemented in the current community standard package (e.g. see l.106 of [this code](https://github.com/achael/eht-imaging/blob/main/ehtim/scattering/stochastic_optics.py) in eht-imaging library). This is a reference implementation of Psaltis et al. 2018. We added the clarification in the source code (commit: https://github.com/EHTJulia/ScatteringOptics.jl/commit/a619d553c22a40d8092943bb7385d0ae932fdfac).
