### 2024-10-16 EHTJulia/ScatteringOptics.jl#22 @kazuakiyama
> A related question about kzetafinder_equation in src/kzetafinders/dipole.jl and src/kzetafinders/vonmises.jl , shouldn't the square difference be penalized instead of value - A^2?

This came from the definition of `NonlinearProblem` in `NonlinearSolve.jl` used in the package.  NonlinearProblem needs the actual equation to be solved rather than the cost function (e.g. squared differences) for the optimization. We tried other optimization packages available in Julia (e.g. `Optim.jl` and `Optimization.jl`), and we found `NonlinearSolve` was the fastest among the packages we tried. This leads to the current implementation choice and definition of the equation.
