+++ dev/assets/tutorials_gettingstarted.md.BhNEQzYx.lean.js
@@ -1,4 +1,4 @@
-import{_ as i,c as a,a5 as n,o as t}from"./chunks/framework.DWm5V20y.js";const e="/ScatteringOptics.jl/dev/assets/cwwbioh.CBf3Exb6.png",l="/ScatteringOptics.jl/dev/assets/uzgcpdn.nmzMopq_.png",h="/ScatteringOptics.jl/dev/assets/cmivldc.BIZLNxDz.png",p="/ScatteringOptics.jl/dev/assets/plrhzcx.DnNIoQj5.png",k="/ScatteringOptics.jl/dev/assets/rdlqwzf.q-9cY4Km.png",F=JSON.parse('{"title":"Getting Started","description":"","frontmatter":{},"headers":[],"relativePath":"tutorials/gettingstarted.md","filePath":"tutorials/gettingstarted.md","lastUpdated":null}'),r={name:"tutorials/gettingstarted.md"};function d(g,s,E,o,c,y){return t(),a("div",null,s[0]||(s[0]=[n(`<h1 id="Getting-Started" tabindex="-1">Getting Started <a class="header-anchor" href="#Getting-Started" aria-label="Permalink to &quot;Getting Started {#Getting-Started}&quot;">​</a></h1><p><code>ScatteringOptics.jl</code> is designed as a package inside the ecosystem of <a href="https://github.com/ptiede/Comrade.jl" target="_blank" rel="noreferrer"><code>Comrade.jl</code></a>. The scattering model in the package can scatter any sky model types from <code>Comrade.jl</code>.</p><h2 id="Loading-your-image" tabindex="-1">Loading your image <a class="header-anchor" href="#Loading-your-image" aria-label="Permalink to &quot;Loading your image {#Loading-your-image}&quot;">​</a></h2><p>Here, we use an example image in <a href="https://github.com/achael/eht-imaging" target="_blank" rel="noreferrer"><code>eht-imaging</code></a>. Data can be downloaded from <a href="./tutorials/data/jason_mad_eofn.fits">here</a>. This is a general relativistic magnetohydrodynamic (GRMHD) model of the magnetic arrestic disk originally from <a href="https://ui.adsabs.harvard.edu/abs/2020MNRAS.494.4168D/abstract" target="_blank" rel="noreferrer">Dexter et al. 2020</a>.</p><div class="language-julia vp-adaptive-theme"><button title="Copy Code" class="copy"></button><span class="lang">julia</span><pre class="shiki shiki-themes github-light github-dark vp-code" tabindex="0"><code><span class="line"><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">using</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;"> CairoMakie</span></span>
[…]
