### 2024-10-16 a2efc8b
add a quick shortcut function to initialize the RefractivePhaseScreen
