### 2024-10-16 f2e0c17
fix another typo in `refractivephasescreen` and updated related documentation.
