### 2024-10-16 b0e0dbd
add quick shortcuts to compute average and emsemble average images.
