+++ src/scatteringmodels/phasescreens/refractivemodel.jl
@@ -50,12 +50,12 @@ average image.
 - `im <: IntensityMap`
 - `Vx_km_per_s` and `Vy_km_per_s` are optional for moving phase screen. 
 """
-function refractivephasescreen(sm::S, im::IntensityMap, Vx_km_per_s::T=0.0, Vy_km_per_s::T=0.0) where {S, T}
-    nx, ny = size(im)[1:2]
-    dx = im.X[2]-im.X[1]  # pixel size in radians
-    dy = im.Y[2]-im.Y[1]
+function refractivephasescreen(sm::S, imap::IntensityMap, Vx_km_per_s::T=0.0, Vy_km_per_s::T=0.0) where {S, T}
+    nx, ny = size(imap)[1:2]
+    dx = imap.X[2]-imap.X[1]  # pixel size in radians
+    dy = imap.Y[2]-imap.Y[1]
 
-    U = typeof(im[1])
+    U = typeof(imap[1])
     UeqT = U == T
     Vx_km = UeqT ? Vx_km_per_s : U(Vx_km_per_s)
     Vy_km = UeqT ? Vy_km_per_s : U(Vy_km_per_s)
@@ -71,10 +71,10 @@ Generates a refractive phase screen, `ϕ`, using StationaryRandomFields.jl the p
 The fourier space 2D noise_screen (defaults to gaussian noise screen if not given) is scaled by the power law,
 `Q`, defined in input AbstractPhaseScreen `psm`. The observing wavelength, `λ_cm`, must be given.
 """
-@inline function phase_screen(psm::AbstractPhaseScreen, λ_cm::Number; noise_screen=nothing)
+@inline function phase_screen(psm::AbstractPhaseScreen, λ_cm::Number; noise_screen=nothing, rng = Random.default_rng())
     # generate noise screen if not provided
     if isnothing(noise_screen)
-        noise_screen = generate_gaussian_noise(psm)
+        noise_screen = generate_gaussian_noise(psm; rng=rng)
     end
 
     # generate phase screen from power law and noise screen
@@ -121,46 +121,79 @@ Returns Fresnel scale corresponding to the given AvstractPhaseScreen object and 
 end
 
 """
-    image_scatter(psm::RefractivePhaseScreen, imap, λ_cm::Number; νref::Number = c_cgs)
+    scatter(psm::RefractivePhaseScreen, imap::IntensityMap; νref::Number = c_cgs, noise_screen=nothing)
 
 Implements full ISM scattering on an unscattered Comrade skymodel intensity map (`imap`). Diffrective blurring and 
 refractive phase screen generation are specific to the scattering parameters defined in the AbstractPhaseScreen
-model `psm`. The observing wavelength `λ_cm` is required.
+model `psm`.
 """
-@inline function image_scatter(psm::AbstractPhaseScreen, imap, λ_cm::Number; νref::Number = c_cgs, noise_screen=nothing)
+@inline function scatter(
+        psm::AbstractPhaseScreen,
+        imap::IntensityMap;
+        νref::Number = c_cgs,
+        noise_screen=nothing,
+        rng = Random.default_rng())
+
     # generate noise screen if not provided
     if isnothing(noise_screen)
-        noise_screen = generate_gaussian_noise(psm)
+        noise_screen = generate_gaussian_noise(psm; rng=rng)
     end
+   
+    # check if imap has a frequncy or time dimension
+    if ndims(imap) > 2
+        throw("The funciton doesn't support multi-dimensional images")
+    end
+
[…]
