+++ docs/src/tutorials/gettingstarted.md
@@ -15,7 +15,7 @@ using VLBISkyModels
 # Alternatively, you can import Comrade.jl instead. Either works.
 # using Comrade
 
-# Load a model image FITS file
+# Load a image model from an image FITS file
 im = load_fits("data/jason_mad_eofn.fits", IntensityMap)
 
 # Plot source image
@@ -43,7 +43,7 @@ print("Frequency of the image: ", νref/1e9," GHz")
 skm = kernelmodel(sm, νref=νref)
 ```
 
-`skm` will be a subtype of `ComradeBase.AbstractModel`, and you can use any methods defined for sky models in `Comrade.jl''s ecosystem. For instance, you can compute the ensemble-average (or diffractively scattered) image using the `convolve` method,
+`skm` is a subtype of `ComradeBase.AbstractModel`, and you can use any methods defined for sky models in `Comrade.jl`'s ecosystem. For instance, you can compute the ensemble-average (or diffractively scattered) image using the `convolve` method from `VLBISkyModels.jl`,
 
 ```@example 1
 # scatter the image
@@ -51,6 +51,27 @@ im_ea = convolve(im, skm)
 
 # Plot source image
 imageviz(im_ea, size=(600, 500), colormap=:afmhot)
+```
+
+You can also directly convolve a skymodel from `ComradeBase.AbstractModel` and implemented in `VLBISkyModels.jl`.
+
+```@example 1
+# Gaussian model from VLBISkyModels.jl
+g = stretched(Gaussian(), μas2rad(10.0), μas2rad(10.0))
+im_g = intensitymap(g, imagepixels(μas2rad(50.0), μas2rad(50.0), 256, 256))
+
+# Plot source image
+imageviz(im_g, size=(600, 500), colormap=:afmhot)
+```
+
+Scattered model and image can be generated by,
+
+```@example 1
+g_ea = convolved(g, skm)
+im_gea = intensitymap(g_ea, imagepixels(μas2rad(50.0), μas2rad(50.0), 256, 256))
+
+# Plot source image
+imageviz(im_gea, size=(600, 500), colormap=:afmhot)
 ```
 
 You can also plot the scattering kernel. For instance in the image domain,
@@ -63,7 +84,7 @@ im_skm = intensitymap(skm, imagepixels(μas2rad(50.0), μas2rad(50.0), 256, 256)
 imageviz(im_skm, size=(600, 500), colormap=:afmhot)
 ```
 
-You can compute the kernel visibility using `visibility_point` method. For instance,
+You can compute the kernel visibility using `visibility_point` method from `VLBISkyModels.jl`. For instance,
 
 ```@example 1
 # computing kernels from 0 to 10 Glambda along RA
@@ -80,14 +101,29 @@ plot!(ax, u/1e9, abs.(vis))
 f
 ```
 
-## Simulating Refractive Scattering
-Otherwise, the scattering kernel `skm` is not required to obtain a fully scattered image. A `RefractivePhaseScreen` type object may be initialized from the `ScatteringModel` and image dimensions. 
+The above tutorial is intentionally written in a low level. There is [emsembleaverage](@ref) method to do a quick shortcut by bypassing the kernel generation.
 
 ```@example 1
-# Create a refractive phase screen model from scattering model and image
+# scatter the image
+im_ea_2 = emsembleaverage(skm, im)
+
+# Plot source image
+imageviz(im_ea_2, size=(600, 500), colormap=:afmhot)
+```
+
+Although this is handy, it may have an extra overhead to initialize `skm` which may slow down highly iterative processes. 
+[emsembleaverage](@ref) method also supports more general skymodels in `ComradeBase.AbstractModel` as an input instead of the image model.
+
+## Simulating Refractive Scattering
[…]
