+++ src/scatteringmodels/abstractscatteringmodel.jl
@@ -157,14 +157,14 @@ end
 """
     emsembleaverage(sm::AbstractScatteringModel, skymodel::AbstractModel, νmodel)
 """
-@inline funciton emsembleaverage(sm::AbstractScatteringModel, skymodel::AbstractModel, νmodel=c_cgs)
+@inline function emsembleaverage(sm::AbstractScatteringModel, skymodel::AbstractModel, νmodel=c_cgs)
     return convolved(skymodel, kernelmodel(sm, νref=νmodel))
 end
 
 """
     emsembleaverage(sm::AbstractScatteringModel, imap::IntensityMap; νref=c_cgs)
 """
-@inline funciton emsembleaverage(sm::AbstractScatteringModel, imap::IntensityMap; νref=c_cgs)
+@inline function emsembleaverage(sm::AbstractScatteringModel, imap::IntensityMap; νref=c_cgs)
     # check if imap has a frequncy or time dimension
     if ndims(imap) > 2
         throw("The funciton doesn't support multi-dimensional images")
