### 2024-10-16 085720c
add metadata in module loading
