### 2024-10-16 5d6687a
rename scatter function to scatter_image
