+++ docs/src/tutorials/diffractive.md
@@ -60,8 +60,8 @@ You can also directly convolve a skymodel from `ComradeBase.AbstractModel` and i
 
 ```@example 1
 # Gaussian model from VLBISkyModels.jl
-g = stretched(Gaussian(), μas2rad(10.0), μas2rad(10.0))
-im_g = intensitymap(g, imagepixels(μas2rad(50.0), μas2rad(50.0), 256, 256))
+g = stretched(Gaussian(), μas2rad(5.0), μas2rad(5.0))
+im_g = intensitymap(g, imagepixels(μas2rad(200.0), μas2rad(200.0), 256, 256))
 
 # Plot source image
 imageviz(im_g, size=(600, 500), colormap=:afmhot)
@@ -71,7 +71,7 @@ Scattered model and image can be generated by,
 
 ```@example 1
 g_ea = convolved(g, skm)
-im_gea = intensitymap(g_ea, imagepixels(μas2rad(50.0), μas2rad(50.0), 256, 256))
+im_gea = intensitymap(g_ea, imagepixels(μas2rad(200.0), μas2rad(200.0), 256, 256))
 
 # Plot source image
 imageviz(im_gea, size=(600, 500), colormap=:afmhot)
@@ -104,6 +104,7 @@ plot!(ax, u/1e9, abs.(vis))
 f
 ```
 
+## A quick shortcut
 The above tutorial is intentionally written in a low level. There is [emsembleaverage](@ref) method to do a quick shortcut by bypassing the kernel generation.
 
 ```@example 1
