+++ dev/assets/chunks/@localSearchIndexroot.Cbwhaiym.js
@@ -1,0 +1,1 @@
+const e='{"documentCount":21,"nextId":21,"documentIds":{"0":"/ScatteringOptics.jl/dev/api#ScateringOptics.jl-API","1":"/ScatteringOptics.jl/dev/api#index","2":"/ScatteringOptics.jl/dev/api#docstrings","3":"/ScatteringOptics.jl/dev/introduction#introduction","4":"/ScatteringOptics.jl/dev/introduction#What-are-interstellar-scattering-and-scintillation?-Why-are-they-important?","5":"/ScatteringOptics.jl/dev/introduction#installation","6":"/ScatteringOptics.jl/dev/math#Brief-Introduction-to-Interstellar-Scattering","7":"/ScatteringOptics.jl/dev/math#Scattering-Regimes-in-Scope","8":"/ScatteringOptics.jl/dev/math#Scattering-Model","9":"/ScatteringOptics.jl/dev/math#references","10":"/ScatteringOptics.jl/dev/tutorials/custommodels#Define-Your-Own-Scattering-Model","11":"/ScatteringOptics.jl/dev/tutorials/custommodels#references","12":"/ScatteringOptics.jl/dev/tutorials/diffractive#Simulate-Diffractive-Scattering","13":"/ScatteringOptics.jl/dev/tutorials/diffractive#Loading-your-image","14":"/ScatteringOptics.jl/dev/tutorials/diffractive#Simulate-diffractive-scattering","15":"/ScatteringOptics.jl/dev/tutorials/diffractive#A-quick-shortcut","16":"/ScatteringOptics.jl/dev/tutorials/nondefaultmodels#Use-Non-default-Models","17":"/ScatteringOptics.jl/dev/tutorials/nondefaultmodels#references","18":"/ScatteringOptics.jl/dev/tutorials/refractive#Simulate-Refractive-Scattering","19":"/ScatteringOptics.jl/dev/tutorials/refractive#Loading-your-image","20":"/ScatteringOptics.jl/dev/tutorials/refractive#Simulating-Refractive-Scattering"},"fieldIds":{"title":0,"titles":1,"text":2},"fieldLength":{"0":[3,1,1],"1":[1,3,42],"2":[1,3,459],"3":[1,1,100],"4":[10,1,128],"5":[1,1,21],"6":[5,1,34],"7":[4,5,168],"8":[2,5,255],"9":[1,5,84],"10":[5,1,254],"11":[1,5,39],"12":[3,1,30],"13":[3,3,136],"14":[3,3,177],"15":[3,3,68],"16":[4,1,95],"17":[1,4,36],"18":[3,1,18],"19":[3,3,78],"20":[3,3,362]},"averageFieldLength":[2.9047619047619047,2.5714285714285716,123.0952380952381],"storedFields":{"0":{"title":"ScateringOptics.jl API","titles":[]},"1":{"title":"Index","titles":["ScateringOptics.jl API"]},"2":{"title":"Docstrings","titles":["ScateringOptics.jl API"]},"3":{"title":"Introduction","titles":[]},"4":{"title":"What are interstellar scattering and scintillation? Why
[…]
