+++ dev/assets/tutorials_diffractive.md.CvP3mtrY.js
@@ -1,4 +1,4 @@
-import{_ as i,c as a,a5 as t,o as n}from"./chunks/framework.DWm5V20y.js";const e="/ScatteringOptics.jl/dev/assets/ieevzdf.CBf3Exb6.png",h="/ScatteringOptics.jl/dev/assets/hwencur.nmzMopq_.png",l="/ScatteringOptics.jl/dev/assets/dxfzaxe.D1qHV6Xh.png",p="/ScatteringOptics.jl/dev/assets/auymmcn.BgHzAqnC.png",k="/ScatteringOptics.jl/dev/assets/pkccajo.BIZLNxDz.png",r="/ScatteringOptics.jl/dev/assets/qexcgbf.DnNIoQj5.png",d="/ScatteringOptics.jl/dev/assets/hwencur.nmzMopq_.png",C=JSON.parse('{"title":"Simulate Diffractive Scattering","description":"","frontmatter":{},"headers":[],"relativePath":"tutorials/diffractive.md","filePath":"tutorials/diffractive.md","lastUpdated":null}'),g={name:"tutorials/diffractive.md"};function E(o,s,c,y,m,F){return n(),a("div",null,s[0]||(s[0]=[t(`<h1 id="Simulate-Diffractive-Scattering" tabindex="-1">Simulate Diffractive Scattering <a class="header-anchor" href="#Simulate-Diffractive-Scattering" aria-label="Permalink to &quot;Simulate Diffractive Scattering {#Simulate-Diffractive-Scattering}&quot;">​</a></h1><p><code>ScatteringOptics.jl</code> is designed as a package inside the ecosystem of <a href="https://github.com/ptiede/Comrade.jl" target="_blank" rel="noreferrer"><code>Comrade.jl</code></a>. The scattering model in the package can scatter any sky model types from <code>Comrade.jl</code>. This page describes how to simulate diffractive scattering.</p><h2 id="Loading-your-image" tabindex="-1">Loading your image <a class="header-anchor" href="#Loading-your-image" aria-label="Permalink to &quot;Loading your image {#Loading-your-image}&quot;">​</a></h2><p>Here, we use an example image in <a href="https://github.com/achael/eht-imaging" target="_blank" rel="noreferrer"><code>eht-imaging</code></a>. Data can be downloaded from <a href="./tutorials/data/jason_mad_eofn.fits">here</a>. This is a general relativistic magnetohydrodynamic (GRMHD) model of the magnetic arrestic disk originally from <a href="https://ui.adsabs.harvard.edu/abs/2020MNRAS.494.4168D/abstract" target="_blank" rel="noreferrer">Dexter et al. 2020</a>.</p><div class="language-julia vp-adaptive-theme"><button title="Copy Code" class="copy"></button><span class="lang">julia</span><pre class="shiki shiki-themes github-light github-dark vp-code" tabindex="0"><code><span class="line"><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">using</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;"> CairoMakie</span></span>
[…]
