### 2024-10-16 EHTJulia/ScatteringOptics.jl#23 @kazuakiyama
@Edenhofer
Based on the comments, tutorials are now dynamically compiled in GitHub flow, rather than showing static outputs.  The documentation now includes different source structures (GRMHD, and a more simple geometric model) as well. We updated the library so that users can use StableRNG for reproducivility and it has been included in the tutorials. This should work as a fairly good end-to-end test as users can directly compare their local outputs with those in the tutorial.

For the basic tests, we will work on them in the next few days --- this is the final piece of the revision as the rest of the comments have been all reflected in the library.
