+++ v0.1.4/assets/index.md._B_sR87P.lean.js
@@ -1,0 +1,1 @@
+import{_ as t,c as e,o as a}from"./chunks/framework.DQADr4iq.js";const p=JSON.parse('{"title":"","description":"","frontmatter":{"layout":"home","hero":{"name":"ScatteringOptics.jl","tagline":"A Julia package to model and simulate interstellar scattering and scintillation","actions":[{"theme":"brand","text":"Introduction","link":"/introduction"},{"theme":"alt","text":"Tutorials","link":"/tutorials/diffractive"},{"theme":"alt","icon":"github","text":"View on Github","link":"https://github.com/EHTJulia/ScatteringOptics.jl"}]}},"headers":[],"relativePath":"index.md","filePath":"index.md","lastUpdated":null}'),i={name:"index.md"};function n(o,r,l,c,s,d){return a(),e("div")}const u=t(i,[["render",n]]);export{p as __pageData,u as default};
