### 2024-10-16 92a6af7
downgrade the version to 0.1.3 as JuliaHub registration of that version failed.
