+++ Project.toml
@@ -1,7 +1,7 @@
 name = "ScatteringOptics"
 uuid = "e317c96f-4a6f-4ae9-8016-294fb9063ff5"
 authors = ["Anna Tartaglia", "Kazunori Akiyama", "Paul Tiede"]
-version = "0.1.4"
+version = "0.1.3"
 
 [deps]
 ComradeBase = "6d8c423b-a35f-4ef1-850c-862fe21f82c4"
@@ -18,7 +18,7 @@ StationaryRandomFields = "6ab34832-f7e5-40a4-9cd7-47ea82b5c144"
 VLBISkyModels = "d6343c73-7174-4e0f-bb64-562643efbeca"
 
 [compat]
-ComradeBase = ">= 0.8"
+ComradeBase = "0.8"
 DocStringExtensions = "0.9"
 EHTUtils = "0.1"
 FFTW = "1"
