### 2024-10-16 85a5ca1
replace tutorials to the latest one in the JOSS paper manuscript
