### 2024-10-16 fe9ade5
add a note about the example data in tutorial
