+++ dev/assets/tutorials_refractive.md.3xEhCBhZ.lean.js
@@ -1,4 +1,4 @@
-import{_ as i,c as a,a5 as e,o as n}from"./chunks/framework.DWm5V20y.js";const t="/ScatteringOptics.jl/dev/assets/aurowtk.CBf3Exb6.png",l="/ScatteringOptics.jl/dev/assets/stksztr.GMBemIl9.png",p="/ScatteringOptics.jl/dev/assets/fqnxsmm.DX0zsh2r.png",h="/ScatteringOptics.jl/dev/assets/dpmpxmj.D1qHV6Xh.png",k="/ScatteringOptics.jl/dev/assets/fysmqhn.Dqm1ccWm.png",u=JSON.parse('{"title":"Simulate Refractive Scattering","description":"","frontmatter":{},"headers":[],"relativePath":"tutorials/refractive.md","filePath":"tutorials/refractive.md","lastUpdated":null}'),r={name:"tutorials/refractive.md"};function d(g,s,c,o,E,m){return n(),a("div",null,s[0]||(s[0]=[e(`<h1 id="Simulate-Refractive-Scattering" tabindex="-1">Simulate Refractive Scattering <a class="header-anchor" href="#Simulate-Refractive-Scattering" aria-label="Permalink to &quot;Simulate Refractive Scattering {#Simulate-Refractive-Scattering}&quot;">​</a></h1><p>Another feature of <code>ScatteringOptics.jl</code> is simulating refractive scattering. This page gives a tutorial to simulate refractive scattering effects.</p><h2 id="Loading-your-image" tabindex="-1">Loading your image <a class="header-anchor" href="#Loading-your-image" aria-label="Permalink to &quot;Loading your image {#Loading-your-image}&quot;">​</a></h2><p>Again, we use an example image in <a href="https://github.com/achael/eht-imaging" target="_blank" rel="noreferrer"><code>eht-imaging</code></a>. Data can be downloaded from <a href="./tutorials/data/jason_mad_eofn.fits">here</a>. This is a general relativistic magnetohydrodynamic (GRMHD) model of the magnetic arrestic disk originally from <a href="https://ui.adsabs.harvard.edu/abs/2020MNRAS.494.4168D/abstract" target="_blank" rel="noreferrer">Dexter et al. 2020</a>.</p><div class="language-julia vp-adaptive-theme"><button title="Copy Code" class="copy"></button><span class="lang">julia</span><pre class="shiki shiki-themes github-light github-dark vp-code" tabindex="0"><code><span class="line"><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">using</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;"> CairoMakie</span></span>
[…]
