### 2024-10-23 827c3ac
Update images benchmarks.md
