+++ docs/make.jl
@@ -27,7 +27,7 @@ makedocs(;
         "Home" => "index.md",
         "Introduction" => "introduction.md",
         "Brief Introduction to interstellar scattering" => "math.md",
-        "Benchmarks" => "benchmarks.md,
+        "Benchmarks" => "benchmarks.md",
         "Tutorials" => TUTORIALS,
         "ScateringOptics.jl API" => "api.md",
     ],
