### 2024-10-24 ccde22b
Add save fits and hdf5 files to docs
