+++ dev/assets/chunks/@localSearchIndexroot.CI6s9nH5.js
@@ -1,0 +1,1 @@
+const e='{"documentCount":26,"nextId":26,"documentIds":{"0":"/ScatteringOptics.jl/dev/api#ScateringOptics.jl-API","1":"/ScatteringOptics.jl/dev/api#index","2":"/ScatteringOptics.jl/dev/api#docstrings","3":"/ScatteringOptics.jl/dev/benchmarks#benchmarks","4":"/ScatteringOptics.jl/dev/benchmarks#speed","5":"/ScatteringOptics.jl/dev/benchmarks#consistency","6":"/ScatteringOptics.jl/dev/introduction#introduction","7":"/ScatteringOptics.jl/dev/introduction#What-are-interstellar-scattering-and-scintillation?-Why-are-they-important?","8":"/ScatteringOptics.jl/dev/introduction#installation","9":"/ScatteringOptics.jl/dev/math#Brief-Introduction-to-Interstellar-Scattering","10":"/ScatteringOptics.jl/dev/math#Scattering-Regimes-in-Scope","11":"/ScatteringOptics.jl/dev/math#Scattering-Model","12":"/ScatteringOptics.jl/dev/math#references","13":"/ScatteringOptics.jl/dev/tutorials/custommodels#Define-Your-Own-Scattering-Model","14":"/ScatteringOptics.jl/dev/tutorials/custommodels#references","15":"/ScatteringOptics.jl/dev/tutorials/diffractive#Simulate-Diffractive-Scattering","16":"/ScatteringOptics.jl/dev/tutorials/diffractive#Loading-your-image","17":"/ScatteringOptics.jl/dev/tutorials/diffractive#Simulate-diffractive-scattering","18":"/ScatteringOptics.jl/dev/tutorials/diffractive#A-quick-shortcut","19":"/ScatteringOptics.jl/dev/tutorials/diffractive#Save-the-tutorial-data","20":"/ScatteringOptics.jl/dev/tutorials/nondefaultmodels#Use-Non-default-Models","21":"/ScatteringOptics.jl/dev/tutorials/nondefaultmodels#references","22":"/ScatteringOptics.jl/dev/tutorials/refractive#Simulate-Refractive-Scattering","23":"/ScatteringOptics.jl/dev/tutorials/refractive#Loading-your-image","24":"/ScatteringOptics.jl/dev/tutorials/refractive#Simulating-Refractive-Scattering","25":"/ScatteringOptics.jl/dev/tutorials/refractive#Save-the-tutorial-data"},"fieldIds":{"title":0,"titles":1,"text":2},"fieldLength":{"0":[3,1,1],"1":[1,3,42],"2":[1,3,459],"3":[1,1,71],"4":[1,1,92],"5":[1,1,88],"6":[1,1,100],"7":[10,1,128],"8":[1,1,21],"9":[5,1,34],"10":[4,5,168],"11":[2,5,255],"12":[1,5,84],"13":[5,1,254],"14":[1,5,39],"15":[3,1,30],"16":[3,3,136],"17":[3,3,177],"18":[3,3,68],"19":[4,3,84],"20":[4,1,95],"21":[1,4,36],"22":[3,1,18],"23":[3,3,78],"24":[3,3,362],"25":[4,3,33]},"averageFieldLength":[2.769230769230769,2.4230
[…]
