### 2024-10-24 53c5d1b
Add assets folder and update docs
