### 2024-10-24 90ae624
put static & dynamically created data into data directory, and manually copy after the build
