### 2024-10-24 b86897c
add manual copy and clean up to fix the issue that data directory will be ignored by DocumenterVitepress.jl
