+++ docs/make.jl
@@ -15,14 +15,33 @@ TUTORIALS = [
     ],
 ]
 
+# Formatter
+format = MarkdownVitepress(
+    repo = "github.com/EHTJulia/ScatteringOptics.jl",
+    clean_md_output = false,
+)
+
+# Check Deploy Decision
+if format.deploy_decision === nothing
+    deploy_config = Documenter.auto_detect_deploy_system()
+    deploy_decision = Documenter.deploy_folder(
+        deploy_config;
+        repo = format.repo, # this must be the full URL!
+        devbranch = format.devbranch,
+        devurl = format.devurl,
+        push_preview=true,
+    )
+else
+    deploy_decision = format.deploy_decision
+end
+
+# build documentation
 makedocs(;
     modules=[ScatteringOptics],
     authors="Anna Tartaglia, Kazunori Akiyama",
     repo="github.com/EHTJulia/ScatteringOptics.jl/blob/{commit}{path}#{line}",
     sitename="ScatteringOptics.jl",
-    format = MarkdownVitepress(
-        repo = "github.com/EHTJulia/ScatteringOptics.jl",
-    ),
+    format = format,
     pages=[
         "Home" => "index.md",
         "Introduction" => "introduction.md",
@@ -33,7 +52,36 @@ makedocs(;
     ],
 )
 
-cp("./build/.documenter/data", "./build/final_site/data", force=true)
+# shortcut to some directories
+builddir = "./build"
+mddir = joinpath(builddir, format.md_output_path)
+fsdir = joinpath(builddir, "final_site")
+
+# manually copy data files into final_site as DocumenterVitepress will ignore files in src/data.
+@info "make.jl: manually copy files in src/data"
+mkpath(joinpath(fsdir,"data"))
+contents = readdir(joinpath(mddir, "data"))
+for item in contents
+    src = joinpath(mddir, "data", item)
+    dst = joinpath(fsdir, "data", item)
+    cp(src, dst)
+end
+
+# manually clean up build 
+clean_md_output = isnothing(format.clean_md_output) ? deploy_decision.all_ok : format.clean_md_output
+if clean_md_output
+    @info "make.jl: manually cleaning up Markdown output."
+    rm(mddir; recursive = true)
+    
+    contents = readdir(fsdir)
+    for item in contents
+        src = joinpath(fsdir, item)
+        dst = joinpath(builddir, item)
+        cp(src, dst)
+    end
+    rm(fsdir; recursive = true)
+    @info "make.jl: Markdown output cleaned up.  Folder looks like:  $(readdir(builddir))"
+end
 
 deploydocs(;
     repo="github.com/EHTJulia/ScatteringOptics.jl",
