### 2024-10-24 2be36e7
add Document object to get the accurate path to the build directory
