### 2024-10-24 47f3c13
fix a bug that the document is not deployed correctly.
