+++ docs/make.jl
@@ -4,6 +4,7 @@ using DocumenterVitepress
 
 DocMeta.setdocmeta!(ScatteringOptics, :DocTestSetup, :(using ScatteringOptics); recursive=true)
 
+# Tutorial files
 TUTORIALS = [
     "Getting Started" => [
         "diffractive.md",
@@ -18,23 +19,10 @@ TUTORIALS = [
 # Formatter
 format = MarkdownVitepress(
     repo = "github.com/EHTJulia/ScatteringOptics.jl",
-    clean_md_output = false,
+    clean_md_output = false,    #  this needs to be false for the manual data copy
 )
 
-# Check Deploy Decision
-if format.deploy_decision === nothing
-    deploy_config = Documenter.auto_detect_deploy_system()
-    deploy_decision = Documenter.deploy_folder(
-        deploy_config;
-        repo = format.repo, # this must be the full URL!
-        devbranch = format.devbranch,
-        devurl = format.devurl,
-        push_preview=true,
-    )
-else
-    deploy_decision = format.deploy_decision
-end
-
+# Keywords for Documenter.jl
 makedocs_kwargs = (
     modules=[ScatteringOptics],
     authors="Anna Tartaglia, Kazunori Akiyama",
@@ -51,12 +39,16 @@ makedocs_kwargs = (
     ],
 )
 
-# build documentation
+# Build documentation
 makedocs(; makedocs_kwargs...)
 
+#---------------------------------------
+# Manually transfer data to be ignored in default by DocumenterVitepress.jl
+#   these are a customized copy of render function in DocumenterVitepress.jl
+
 # shortcut to some directories
-doc = Documenter.Document(; makedocs_kwargs...)
-builddir = isabspath(doc.user.build) ? doc.user.build : joinpath(doc.user.root, doc.user.build)
+doc = Documenter.Document(; makedocs_kwargs...) # snipet from makedocs.jl in Documenter.jl
+builddir = isabspath(doc.user.build) ? doc.user.build : joinpath(doc.user.root, doc.user.build) 
 mddir = joinpath(builddir, format.md_output_path)
 fsdir = joinpath(builddir, "final_site")
 
@@ -70,9 +62,22 @@ for item in contents
     cp(src, dst)
 end
 
-# manually clean up build 
-clean_md_output = deploy_decision.all_ok
-if clean_md_output
+# check if the document will be deployed
+if format.deploy_decision === nothing
+    deploy_config = Documenter.auto_detect_deploy_system()
+    deploy_decision = Documenter.deploy_folder(
+        deploy_config;
+        repo = format.repo, # this must be the full URL!
+        devbranch = format.devbranch,
+        devurl = format.devurl,
+        push_preview=true,
+    )
+else
+    deploy_decision = format.deploy_decision
+end
+
+# manually clean up the build directory if the documentation to be deployed.
+if deploy_decision.all_ok
     @info "make.jl: manually cleaning up Markdown output."
     rm(mddir; recursive = true)
     
@@ -85,7 +90,9 @@ if clean_md_output
     rm(fsdir; recursive = true)
     @info "make.jl: Markdown output cleaned up.  Folder looks like:  $(readdir(builddir))"
 end
+#---------------------------------------
 
+# Deploy Documentation
 deploydocs(;
     repo="github.com/EHTJulia/ScatteringOptics.jl",
     target = "build",
