### 2024-10-25 463c801
add warnings for links
