+++ dev/assets/chunks/@localSearchIndexroot.mzX19PWK.js
@@ -1,0 +1,1 @@
+const e='{"documentCount":26,"nextId":26,"documentIds":{"0":"/ScatteringOptics.jl/dev/api#ScateringOptics.jl-API","1":"/ScatteringOptics.jl/dev/api#index","2":"/ScatteringOptics.jl/dev/api#docstrings","3":"/ScatteringOptics.jl/dev/benchmarks#benchmarks","4":"/ScatteringOptics.jl/dev/benchmarks#speed","5":"/ScatteringOptics.jl/dev/benchmarks#consistency","6":"/ScatteringOptics.jl/dev/custommodels#Define-Your-Own-Scattering-Model","7":"/ScatteringOptics.jl/dev/custommodels#references","8":"/ScatteringOptics.jl/dev/diffractive#Simulate-Diffractive-Scattering","9":"/ScatteringOptics.jl/dev/diffractive#Loading-your-image","10":"/ScatteringOptics.jl/dev/diffractive#Simulate-diffractive-scattering","11":"/ScatteringOptics.jl/dev/diffractive#A-quick-shortcut","12":"/ScatteringOptics.jl/dev/diffractive#Save-the-tutorial-data","13":"/ScatteringOptics.jl/dev/introduction#introduction","14":"/ScatteringOptics.jl/dev/introduction#What-are-interstellar-scattering-and-scintillation?-Why-are-they-important?","15":"/ScatteringOptics.jl/dev/introduction#installation","16":"/ScatteringOptics.jl/dev/math#Brief-Introduction-to-Interstellar-Scattering","17":"/ScatteringOptics.jl/dev/math#Scattering-Regimes-in-Scope","18":"/ScatteringOptics.jl/dev/math#Scattering-Model","19":"/ScatteringOptics.jl/dev/math#references","20":"/ScatteringOptics.jl/dev/nondefaultmodels#Use-Non-default-Models","21":"/ScatteringOptics.jl/dev/nondefaultmodels#references","22":"/ScatteringOptics.jl/dev/refractive#Simulate-Refractive-Scattering","23":"/ScatteringOptics.jl/dev/refractive#Loading-your-image","24":"/ScatteringOptics.jl/dev/refractive#Simulating-Refractive-Scattering","25":"/ScatteringOptics.jl/dev/refractive#Save-the-tutorial-data"},"fieldIds":{"title":0,"titles":1,"text":2},"fieldLength":{"0":[3,1,1],"1":[1,3,42],"2":[1,3,459],"3":[1,1,71],"4":[1,1,82],"5":[1,1,74],"6":[5,1,254],"7":[1,5,39],"8":[3,1,30],"9":[3,3,144],"10":[3,3,177],"11":[3,3,68],"12":[4,3,101],"13":[1,1,100],"14":[10,1,128],"15":[1,1,21],"16":[5,1,34],"17":[4,5,168],"18":[2,5,255],"19":[1,5,84],"20":[4,1,95],"21":[1,4,36],"22":[3,1,18],"23":[3,3,87],"24":[3,3,362],"25":[4,3,46]},"averageFieldLength":[2.769230769230769,2.4230769230769234,114.46153846153847],"storedFields":{"0":{"title":"ScateringOptics.jl API","titles":[]},"1":{"title
[…]
