+++ dev/assets/benchmarks.md.xHWXQ9Xw.js
@@ -1,0 +1,1 @@
+import{_ as i,c as a,a5 as n,j as t,a as s,o}from"./chunks/framework.DWm5V20y.js";const l="/ScatteringOptics.jl/dev/assets/speed_dipole.0zob2s6s.png",r="/ScatteringOptics.jl/dev/assets/speed_boxcar.DNaYty9n.png",m="/ScatteringOptics.jl/dev/assets/speed_vonmises.QMxRx1fN.png",d="/ScatteringOptics.jl/dev/assets/kernel87.B3OWRtHj.png",p="/ScatteringOptics.jl/dev/assets/kernel36.BqpikCNz.png",y=JSON.parse('{"title":"Benchmarks","description":"","frontmatter":{},"headers":[],"relativePath":"benchmarks.md","filePath":"benchmarks.md","lastUpdated":null}'),c={name:"benchmarks.md"},h={class:"MathJax",jax:"SVG",style:{direction:"ltr",position:"relative"}},g={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-0.05ex"},xmlns:"http://www.w3.org/2000/svg",width:"4.495ex",height:"2.005ex",role:"img",focusable:"false",viewBox:"0 -864 1986.7 886","aria-hidden":"true"},u={class:"MathJax",jax:"SVG",style:{direction:"ltr",position:"relative"}},Q={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-0.05ex"},xmlns:"http://www.w3.org/2000/svg",width:"4.495ex",height:"2.02ex",role:"img",focusable:"false",viewBox:"0 -871.1 1986.7 893.1","aria-hidden":"true"};function T(x,e,f,w,k,v){return o(),a("div",null,[e[9]||(e[9]=n('<h1 id="benchmarks" tabindex="-1">Benchmarks <a class="header-anchor" href="#benchmarks" aria-label="Permalink to &quot;Benchmarks&quot;">​</a></h1><p>The Julia implementation of the EHT scattering framework takes advantage of Julia&#39;s fast Auto-Differentiation (AD) capabilities, which accelerate the simulation of interstellar scattering. As the overall imaging process can consist of thousands to tens of thousands of iterations in order to optimize scattering parameters and the final output image, even a small imprdovements in simulation speed can result in significant cuts in runtime. It is then crucial that our implementation performs faster than the existing python framework, <code>eht-imaging</code>, and produces consistent results. Here, we present benchmarks between the two implementations.</p><h2 id="speed" tabindex="-1">Speed <a class="header-anchor" href="#speed" aria-label="Permalink to &quot;Speed&quot;">​</a></h2><p>With Julia&#39;s speed improvements, the <code>ScatteringOptics.jl</code> scattering kernel loads 100 times faster than the <code>eht-imaging</code> kernel. The scattering kernel may be called upon to compute visibilities for a given sample of Fourier space points. In these computations, <code>ScatteringOptics.jl</code> also exhibits speed improvements up to 100 times over the three included models (Dipole, Boxcar, and Von Mises; see <a href="/ScatteringOptics.jl/dev/nondefaultmodels#Use-Non-default-Models">Use Non-default Models</a> for more info) and varying observing wavelengths. These computation runtimes are compared over a range of sample in the plots below.</p><p><img src="'+l+'" alt=""><
[…]

[…]
