### 2024-10-25 d1d2cb6
fix a bug that phase_screen ignores the input noise screen
