### 2024-10-25 bb02992
Change image layout
