### 2024-10-25 966f729
Update diffractive.md
