+++ v0.1.4/assets/nondefaultmodels.md.DPQnEhY6.js
@@ -1,0 +1,3 @@
+import{_ as r,c as d,a5 as l,G as s,w as t,B as n,o as p,j as e}from"./chunks/framework.BHIid8ey.js";const v=JSON.parse('{"title":"Use Non-default Models","description":"","frontmatter":{},"headers":[],"relativePath":"nondefaultmodels.md","filePath":"nondefaultmodels.md","lastUpdated":null}'),h={name:"nondefaultmodels.md"};function c(u,a,g,k,b,f){const i=n("PluginTabsTab"),o=n("PluginTabs");return p(),d("div",null,[a[3]||(a[3]=l(`<h1 id="Use-Non-default-Models" tabindex="-1">Use Non-default Models <a class="header-anchor" href="#Use-Non-default-Models" aria-label="Permalink to &quot;Use Non-default Models {#Use-Non-default-Models}&quot;">​</a></h1><p>As noted in <a href="/ScatteringOptics.jl/v0.1.4/math#Brief-Introduction-to-Interstellar-Scattering">Brief Introduction to Interstellar Scattering</a>, the default scattering model is the <em>dipole</em> model proposed in Johnson &amp; Narayan 2016 ([1]). This is well consistent with actual interstellar scattering effects observed for the Galactic Center Sgr A* and other compact sources ([2]). Indeed you can confirm by</p><div class="language-julia vp-adaptive-theme"><button title="Copy Code" class="copy"></button><span class="lang">julia</span><pre class="shiki shiki-themes github-light github-dark vp-code" tabindex="0"><code><span class="line"><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">using</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;"> ScatteringOptics</span></span>
+<span class="line"></span>
[…]
