### 2024-10-25 EHTJulia/ScatteringOptics.jl#23 @annatartaglia
@Edenhofer
After the discussion within the team, we have decided to additionally provide the reference processed data in our documentation, which will provide the ultimate end-to-end unit tests to the users. You can find the corresponding documentation pages here:

- https://ehtjulia.github.io/ScatteringOptics.jl/v0.1.4/diffractive
- https://ehtjulia.github.io/ScatteringOptics.jl/v0.1.4/refractive

As @kazuakiyama wrote, the julia code blocks are executed upon each commit to GitHub, which will automatically keep updating the outputs with respect to the updates in the package. In addition to the plots that @kazuakiyama mentioned, the documentation further produces the corresponding image data and some numerical array for the scattering kernel in the community standard formats FITS and HDF5, respectively. We provide links to those data in the tutorial pages, where users can download and compare with their own outputs. We now include two different types of input models in the tutorial (one provided fits file and one Gaussian model that users can generate themselves using VLBISkyModels.jl), which will also help user tests. Thanks for the helpful suggestions!
