### 2024-10-25 2b333b7
edit the triggering criteria to limit the compaile of PDF only to main branch
