### 2024-10-25 #6354 @annatartaglia
@dfm

On behalf of the authors, I resubmit our JOSS paper here. We have fully revised the package, its documentation, and our JOSS manuscript according to many helpful and constructive comments made by @Edenhofer and @tomkimpson. The package also has other major updates to support the latest Comrade.jl’s ecosystem, which is not related to the previous review.

You can find the revised paper PDF [here](https://github.com/EHTJulia/ScatteringOptics.jl/blob/joss_v2/paper/paper.pdf). The revised documentation can be found [here](https://ehtjulia.github.io/ScatteringOptics.jl/v0.1.4/). The corresponding package can be found in the tag [joss_v2](https://github.com/EHTJulia/ScatteringOptics.jl/tree/joss_v2) of the branch `joss-review`, which is equivalent to the package [`v0.1.4`](https://github.com/EHTJulia/ScatteringOptics.jl/releases/tag/v0.1.4) of the `main` branch.

For your reference, you can compare them with the documentation at the time of the first submission ([here](https://ehtjulia.github.io/ScatteringOptics.jl/v0.1.2/)) as well as the package with the tag [`joss_v1`](https://github.com/EHTJulia/ScatteringOptics.jl/tree/joss_v1).

In the following replies, we will describe our point-to-point answer to the referee comments by each reviewer.

We believe the paper and package now meet the standard of the JOSS. Please let us know if you have any questions or suggestions for our revision and replies below.

<<Perhaps break the replies here, and make a separated reply to each of referees?>>

***Responses to the report by @Edenhofer***

> The paper is well written, summarizes the need for the software well, and focuses on the essential mathematics.

> However, as a non-expert in the field of radio astronomy, I think that the mathematics discussed should be more closely tied to what ScatteringOptics.jl provides.

We are, unfortunately, limited by the number of words allowed for a JOSS paper (250-1000). Because the mathematics used in our package are so extensive, it is not possible to include all of it in the paper. In the revision, instead, we have significantly extended the texts in our documentation. See ``Brief Introduction to Interstellar Scattering’’ in the updated documentation.

 > In particular, it is not clear to me how the equation after l70 and the equation after l74 are related. I probably am missing something, but I thought the scattering is performed in image space and only and are relevant for ScatteringOptics.jl.

To clarify your initial question–the l74 equation is the l70 equation represented in Fourier “visibility” space. The exponential term in l74 is the Fourier space transformation of our scattering kernel ‘G’. Scattering is easier to handle in the Fourier domain due to the nature of Radio Interferometric measurements. The kernel is well described in this space. We have added a line clarifying this (l71).

This is also now clarified in ``Brief Introduction to Interstellar Scattering’’ in the updated documentation.

> If ScatteringOptics.jl only returns the effective scattering in visibility space, please state this more clearly, e.g. in l82ff.

This is correct. ScatteringOptics.jl produces a sky model of Comrade.jl for the scattering kernel, described analytically in Fourier space. Comrade.jl provides a powerful tool kit to compute the kernel on the image domain through FFT, and blur arbitral sky models with this kernel. For instance, this allows users to see the kernel also in the image domain using the standard Comrade.jl’s function.

> l65 Why does ϕ depend twice on r?

Thanks for pointing this out. Indeed, there is no strong reason to put r on \phi. \phi_r(r) has been replaced to \phi(r) in the paper.

> l65-l67 What is the difference between the phase screen and the spatial structure function of the phase screen?

The structure function is a measure of second-order phase changes in the phase screen at different locations. A complete description of the mathematics involved in calculating the structure function can be found in this paper: https://arxiv.org/abs/1805.01242. Due to constraints on the length of our paper, it is not possible to detail all of the mathematics of scattering. We have added further explanation of this function in the documentation ``Brief Introduction to Interstellar Scattering’’.

> l70 How is G related to ϕ?

Our previous clarifications have answered this question — the scattering kernel relates to the structure function in Fourier space, and the structure function itself depends on \phi.  It is also now explained in the documentation. Please see ``Brief Introduction to Interstellar Scattering’’.

> r is defined in l71 and again in l65; please merge the definitions.

The definition is now in l65.

> l76 It is not clear to me what "earth-screen distance D" refers to; please rephrase.

This refers to the distance between the earth and the scattering material. We have added that clarification.

> Is eq. following l74 an alternative formulation of l70 in visibility space?

Yes, refer to the above notes.

> l97 It would be great if the authors could provide an example that is reproducible from scratch, e.g., by providing "jason_mad_eofn.fits".

This FITS file is now included in the tutorial. Also, the tutorial now fixed the random number generator so that the results will be reproducible.

> l102 Wrap line.

This is fixed now.

> Figure 1 and Figure 2: Increase DPI for publication?

This is fixed as well, thanks!

***Response to the issues posted by @Edenhofer***

We made public responses to each of the issues posted by @Edenhofer. See below.
https://github.com/EHTJulia/ScatteringOptics.jl/issues/22
https://github.com/EHTJulia/ScatteringOptics.jl/issues/23

***Response to the referee report by @tomkimpson***

> In the manuscript, is it really true that current scattering models are written in Python? If so, then it is fine to refer to the speed of Julia as the motivation. If not (I always thought these things were done in C++, but I could just be ignorant), then perhaps the composability with the wider EHTJulia organisation is a more accurate justification.

Yes, this is true. The current Python and this new Julia implementation are both based on the theoretical discovery and framework by Blandford & Narayan 1985 (MNRAS, 213, 591) and later modernized by Johnson & Narayan 2016 (ApJ, 826, 170), which has been the community standard for almost a decade.

When the measurements are obtained within a timescale longer than the diffractive timescale (often seconds or shorter), the refractive scattering effects can be accurately estimated through a simplified, semi-analytic geometrical optics framework. This works for the vast majority of radio interferometric observations.

This allows a significant acceleration to simulate refractive scattering effects compared to full numerical models (e.g. Johnson & Gwinn 2015, ApJ, 805, 180), which are much slower and in addition have an issue with numerical accuracy. We didn’t mention it because the full numerical model of Johnson & Gwinn 2015 has not been publicly released and was quickly replaced within a year by the framework of this package.

> In the manuscript, it claims that ScatteringOptics.jl includes a set of abstract types that enable users to define other phase screen models. It is not obvious to me how a user would define some other phase screen model - perhaps a MWE would be useful.

We implemented all models in the literature, but yes it is indeed straightforward for interested users to implement their own scattering models. We have added a new documentation page ``Define Your Own Scattering Model’’.

> The MWE in the tutorial is difficult for a user to reproduce without the accompanying .fits file. If sufficiently small, it could be worth including jason_mad_eofn.fits in a data/ directory.

This FITS file is now included in the tutorial. Also, the tutorial now fixed the random number generator so that the results will be reproducible.

> I would really really like to see some unit tests. I see that @Edenhofer has already opened an issue about this. This would make is easier to verify the functionality of the software, and have confidence that everything is working as expected.

See the answer to the GitHub issue ([here](https://github.com/EHTJulia/ScatteringOptics.jl/issues/23)).

> Some community guidelines should be included for third parties who want to contribute or raise issues

We have added the [ColPrac](https://github.com/SciML/ColPrac) guideline to the readme page.
