### 2024-10-26 EHTJulia/ScatteringOptics.jl#23 @Edenhofer
This is great! The documentation is now much more accessible and I can reproduce the demos without issues.

One final remark:  I think it is fantastic that the code within the documentation is automatically run as part of the CI but I think it would be great if the essential parts therein would also be proper unit tests such that errors can be flagged automatically.
