### 2024-10-26 #6354 @Edenhofer
I think the submission is in very good shape. I have two remaining points that I would like to bring up before completing my review:
* I like the benchmarks comparing the Julia implementation versus the old python implementation. Ideally I would like to reproduce these. Could you share the script to reproduce the plots shown at https://ehtjulia.github.io/ScatteringOptics.jl/v0.1.4/benchmarks .
* I think it would be great to have test that can be run locally and fail if something in the software goes awfully wrong. Running the code in the documentation live is great but it is easy to miss errors in the documentation.
