### 2024-10-26 EHTJulia/ScatteringOptics.jl#22 @Edenhofer
Thank you for your detailed reply! This all makes sense :)
