### 2024-10-26 c672147
set up the enviroment for runtest
