+++ test/runtests.jl
@@ -1,5 +1,9 @@
+using BenchmarkTools
+using HDF5
 using ScatteringOptics
+using StableRNGs
 using Test
+using VLBISkyModels
 
 @testset "ScatteringOptics.jl" begin
     # Write your tests here.
