+++ test/Project.toml
@@ -1,0 +1,7 @@
+[deps]
+BenchmarkTools = "6e4b80f9-dd63-53aa-95a3-0cdb28fa8baf"
+HDF5 = "f67ccb44-e63f-5c2f-98bd-6dc0ccc4ba2f"
+ScatteringOptics = "e317c96f-4a6f-4ae9-8016-294fb9063ff5"
+StableRNGs = "860ef19b-820b-49d6-a774-d7a799459cd3"
+Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+VLBISkyModels = "d6343c73-7174-4e0f-bb64-562643efbeca"
