+++ dev/assets/refractive.md.jnzRKJbT.lean.js
@@ -1,4 +1,4 @@
-import{_ as l,c as a,a5 as e,j as s,a as t,o as n}from"./chunks/framework.BMbWnoF-.js";const p="/ScatteringOptics.jl/dev/assets/khcdbuy.CBf3Exb6.png",h="/ScatteringOptics.jl/dev/assets/tiacbmp.9QsyFGId.png",r="/ScatteringOptics.jl/dev/assets/rlrpgme.BXZtWIg7.png",k="/ScatteringOptics.jl/dev/assets/aqczkfv.D1qHV6Xh.png",d="/ScatteringOptics.jl/dev/assets/tqioxqd.Cm55C9do.png",b=JSON.parse('{"title":"Simulate Refractive Scattering","description":"","frontmatter":{},"headers":[],"relativePath":"refractive.md","filePath":"refractive.md","lastUpdated":null}'),o={name:"refractive.md"},g={class:"MathJax",jax:"SVG",style:{direction:"ltr",position:"relative"}},c={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-0.05ex"},xmlns:"http://www.w3.org/2000/svg",width:"6.883ex",height:"2.005ex",role:"img",focusable:"false",viewBox:"0 -864 3042.5 886","aria-hidden":"true"},m={class:"MathJax",jax:"SVG",style:{direction:"ltr",position:"relative"}},E={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-0.05ex"},xmlns:"http://www.w3.org/2000/svg",width:"7.683ex",height:"2.005ex",role:"img",focusable:"false",viewBox:"0 -864 3396 886","aria-hidden":"true"};function y(u,i,f,F,T,v){return n(),a("div",null,[i[7]||(i[7]=e(`<h1 id="Simulate-Refractive-Scattering" tabindex="-1">Simulate Refractive Scattering <a class="header-anchor" href="#Simulate-Refractive-Scattering" aria-label="Permalink to &quot;Simulate Refractive Scattering {#Simulate-Refractive-Scattering}&quot;">​</a></h1><p>Another feature of <code>ScatteringOptics.jl</code> is simulating refractive scattering. This page gives a tutorial to simulate refractive scattering effects.</p><h2 id="Loading-your-image" tabindex="-1">Loading your image <a class="header-anchor" href="#Loading-your-image" aria-label="Permalink to &quot;Loading your image {#Loading-your-image}&quot;">​</a></h2><p>Again, we use an example image in <a href="https://github.com/achael/eht-imaging" target="_blank" rel="noreferrer"><code>eht-imaging</code></a>. Data can be downloaded from <a href="./data/jason_mad_eofn.fits">here</a> (please open in a new window. otherwise you will get 404 error). This is a general relativistic magnetohydrodynamic (GRMHD) model of the magnetic arrestic disk originally from <a href="https://ui.adsabs.harvard.edu/abs/2014IAUS..303..298D" target="_blank" rel="noreferrer">Dexter et al. 2014</a>.</p><div class="language-julia vp-adaptive-theme"><button title="Copy Code" class="copy"></button><span class="lang">julia</span><pre class="shiki shiki-themes github-light github-dark vp-code" tabindex="0"><code><span class="line"><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">using</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;"> CairoMakie</span></span>
[…]

[…]
