### 2024-10-26 48c4ad7
add an option to use the exact scattering kenerl for `scatter_image` funcitons
