### 2024-10-26 02579ac
fix typos in the benchmark page of the documentation
