### 2024-10-30 EHTJulia/ScatteringOptics.jl#23 @kazuakiyama
@Edenhofer
Our decision was in part because this automatic run with CI has a convenient procedure that shows visualization of products as part of the documentation, so we thought it was a more digestible automatic test than more standard run tests. I do agree that implementing run tests is indeed good for reproducibility. Let us discuss what would be the best tests for us. We will likely implement run tests similar to the documentation following your suggestion.
