### 2024-11-01 #6354 @kazuakiyama
@Edenhofer
For the first point, I have a question for you to make sure we revise in a reasonable way. We will attempt to put benchmark scripts in the repository. Our tentative plan is creating the `examples/benchmark` directory, and put scripts and outputs there. Python scripts wouldn't be able to be tested automatically, just because it is outside of our CI framework following Julia's standards. Instead, we will likely put a setup script to reproduce those results. Does that sound right for you?

For the second point, we will explore the best tests to be included as unit tests as we replied to the issue at the repository.
