### 2024-11-02 #6354 @Edenhofer
I am happy with any accessible way to reproduce the benchmark. What you propose sounds great! :)
