### 2024-11-02 EHTJulia/ScatteringOptics.jl#23 @Edenhofer
I think it is totally fine if the documentation and the test share a lot of code. I'm merely advocating for some form of unit tests :)
