### 2024-11-07 #6354 @warrickball
Hi @annatartaglia — I've been processing submissions and noticed that this repo was submitted to JOSS again about two weeks ago even though this review is still technically active. I've rejected the other submission but if there's anything about the process that is confusing, let us know.
