### 2024-11-09 790b948
add tests for diffractive scattering kernel
