### 2024-11-10 91ee989
CompatHelper: bump compat for NonlinearSolve to 4, (keep existing compat)
