+++ Project.toml
@@ -24,7 +24,7 @@ EHTUtils = "0.1"
 FFTW = "1"
 HypergeometricFunctions = "0.3"
 Interpolations = "0.14, 0.15"
-NonlinearSolve = "1, 3"
+NonlinearSolve = "1, 3, 4"
 QuadGK = "2"
 SpecialFunctions = "2"
 StationaryRandomFields = "0.1"
