### 2024-11-13 c08920c
add use_approx options to ensembleaverage methods. docstings are also updated.
