+++ src/scatteringmodels/phasescreens/refractivemodel.jl
@@ -48,7 +48,7 @@ average image.
 
 - `sm <: AbstractScatteringModel`
 - `im <: IntensityMap`
-- `Vx_km_per_s` and `Vy_km_per_s` are optional for moving phase screen. 
+- `Vx_km_per_s` and `Vy_km_per_s` are optional for moving phase screen, which is not yet implemented. 
 """
 function refractivephasescreen(sm::S, imap::IntensityMap, Vx_km_per_s::T=0.0, Vy_km_per_s::T=0.0) where {S, T}
     nx, ny = size(imap)[1:2]
@@ -121,11 +121,17 @@ Returns Fresnel scale corresponding to the given AvstractPhaseScreen object and 
 end
 
 """
-    scatter_image(psm::RefractivePhaseScreen, imap::IntensityMap; νref::Number = c_cgs, noise_screen=nothing)
+    scatter_image(psm::AbstractPhaseScreen, imap::IntensityMap; νref::Number = c_cgs, noise_screen=nothing, rng = Random.default_rng(), use_approx::Bool=true)
 
-Implements full ISM scattering on an unscattered Comrade skymodel intensity map (`imap`). Diffrective blurring and 
-refractive phase screen generation are specific to the scattering parameters defined in the AbstractPhaseScreen
-model `psm`.
+# Arguments
+- `psm::AbstractPhaseScreen`: An instance of the phase screen.
+- `imap::IntensityMap`: An instance of the intensity map.
+
+# Parameters
+- `νref::Number=c_cgs`: The frequency in Hz to be used to simulate scattering effects. If the frequency of the image is not provided, this value is used.
+- `noise_screen=nothing`: The noise screen to be used to generate the phase screen. If not provided, a gaussian noise screen is generated.
+- `rng = Random.default_rng()`: The random number generator.
+- `use_approx::Bool=true`: If true, the approximate analytic formula is used to compute the scattering kernel. Otherwise semi-analytic formula is used.
 """
 @inline function scatter_image(
         psm::AbstractPhaseScreen,
@@ -190,7 +196,21 @@ model `psm`.
 end
 
 """
-    scatter_image(sm::AbstractScatteringModel, imap::IntensityMap, λ_cm::Number; νref::Number = c_cgs, rng = Random.default_rng())
+    scatter_image(sm, imap::IntensityMap; νref::Number = c_cgs, Vx_km_per_s=0.0, Vy_km_per_s=0.0, rng = Random.default_rng(), use_approx::Bool=true)
+
+Simulates the full interstellar scattering on an unscattered Comrade skymodel intensity map (`imap`).
+The frequency of the image, if exists in its metadata, is used to simulate the scattering effects. Otherwise `νref` is used.
+
+# Arguments
+- `sm::AbstractScatteringModel`: An instance of the scattering model.
+- `imap::IntensityMap`: An instance of the intensity map.
+
+# Parameters
+- `νref::Number=c_cgs`: The frequency in Hz to be used to simulate scattering effects. If the frequency of the image is not provided, this value is used.
+- `Vx_km_per_s=0.0`: The velocity of the observer in the x direction in km/s. (just placeholder for future implementation)
+- `Vy_km_per_s=0.0`: The velocity of the observer in the y direction in km/s. (just placeholder for future implementation)
+- `rng = Random.default_rng()`: The random number generator.
+- `use_approx::Bool=true`: If true, the approximate analytic formula is used to compute the scattering kernel. Otherwise semi-analytic formula is used.
 """
 @inline function scatter_image(
     sm::AbstractScatteringModel,
