### 2024-11-13 b640b50
add end-to-end tests comparing with reference outputs
