+++ test/runtests.jl
@@ -1,3 +1,4 @@
+using JLD2
 using ScatteringOptics
 using StableRNGs
 using Statistics
@@ -5,5 +6,6 @@ using Test
 using VLBISkyModels
 
 @testset "ScatteringOptics.jl" begin
-    include("scatteringmodels.jl")
+    include("internalconsistency.jl")
+    include("simtest.jl")
 end
