### 2024-11-13 951a979
add docs and comments. no net changes
