### 2024-11-13 c90b07e
update token command one more time following the community guide
