+++ .github/workflows/CI.yml
@@ -34,8 +34,10 @@ jobs:
       - uses: julia-actions/julia-runtest@v1
       - uses: julia-actions/julia-processcoverage@v1
       - uses: codecov/codecov-action@v4
-        env:
-          CODECOV_TOKEN: ${{ secrets.CODECOV_TOKEN }}
+          with:
+            # possibly other stuff
+            token: ${{ secrets.CODECOV_TOKEN }}
+            fail_ci_if_error: false  # or true if you want CI to fail when Codecov fails
   docs:
     name: Documentation
     runs-on: ubuntu-latest
