+++ .github/workflows/CI.yml
@@ -35,9 +35,8 @@ jobs:
       - uses: julia-actions/julia-processcoverage@v1
       - uses: codecov/codecov-action@v4
           with:
-            # possibly other stuff
             token: ${{ secrets.CODECOV_TOKEN }}
-            fail_ci_if_error: false  # or true if you want CI to fail when Codecov fails
+            fail_ci_if_error: false 
   docs:
     name: Documentation
     runs-on: ubuntu-latest
