+++ .github/workflows/CI.yml
@@ -34,9 +34,9 @@ jobs:
       - uses: julia-actions/julia-runtest@v1
       - uses: julia-actions/julia-processcoverage@v1
       - uses: codecov/codecov-action@v4
-          with:
-            token: ${{ secrets.CODECOV_TOKEN }}
-            fail_ci_if_error: false 
+        with:
+          token: ${{ secrets.CODECOV_TOKEN }}
+          fail_ci_if_error: false 
   docs:
     name: Documentation
     runs-on: ubuntu-latest
