+++ v0.1.5/assets/chunks/@localSearchIndexroot.D7w-PqfW.js
@@ -1,0 +1,1 @@
+const e='{"documentCount":28,"nextId":28,"documentIds":{"0":"/ScatteringOptics.jl/v0.1.5/api#ScateringOptics.jl-API","1":"/ScatteringOptics.jl/v0.1.5/api#index","2":"/ScatteringOptics.jl/v0.1.5/api#docstrings","3":"/ScatteringOptics.jl/v0.1.5/benchmarks#benchmarks","4":"/ScatteringOptics.jl/v0.1.5/benchmarks#speed","5":"/ScatteringOptics.jl/v0.1.5/benchmarks#consistency","6":"/ScatteringOptics.jl/v0.1.5/custommodels#Define-Your-Own-Scattering-Model","7":"/ScatteringOptics.jl/v0.1.5/custommodels#references","8":"/ScatteringOptics.jl/v0.1.5/diffractive#Simulate-Diffractive-Scattering","9":"/ScatteringOptics.jl/v0.1.5/diffractive#Loading-your-image","10":"/ScatteringOptics.jl/v0.1.5/diffractive#Simulate-diffractive-scattering","11":"/ScatteringOptics.jl/v0.1.5/diffractive#A-quick-shortcut","12":"/ScatteringOptics.jl/v0.1.5/diffractive#Save-the-tutorial-data","13":"/ScatteringOptics.jl/v0.1.5/diffractive#Unit-Tests","14":"/ScatteringOptics.jl/v0.1.5/introduction#introduction","15":"/ScatteringOptics.jl/v0.1.5/introduction#What-are-interstellar-scattering-and-scintillation?-Why-are-they-important?","16":"/ScatteringOptics.jl/v0.1.5/introduction#installation","17":"/ScatteringOptics.jl/v0.1.5/math#Brief-Introduction-to-Interstellar-Scattering","18":"/ScatteringOptics.jl/v0.1.5/math#Scattering-Regimes-in-Scope","19":"/ScatteringOptics.jl/v0.1.5/math#Scattering-Model","20":"/ScatteringOptics.jl/v0.1.5/math#references","21":"/ScatteringOptics.jl/v0.1.5/nondefaultmodels#Use-Non-default-Models","22":"/ScatteringOptics.jl/v0.1.5/nondefaultmodels#references","23":"/ScatteringOptics.jl/v0.1.5/refractive#Simulate-Refractive-Scattering","24":"/ScatteringOptics.jl/v0.1.5/refractive#Loading-your-image","25":"/ScatteringOptics.jl/v0.1.5/refractive#Simulating-Refractive-Scattering","26":"/ScatteringOptics.jl/v0.1.5/refractive#Save-the-tutorial-data-and-do-some-unit-tests","27":"/ScatteringOptics.jl/v0.1.5/refractive#Unit-Tests"},"fieldIds":{"title":0,"titles":1,"text":2},"fieldLength":{"0":[3,1,1],"1":[1,3,42],"2":[1,3,478],"3":[1,1,71],"4":[1,1,109],"5":[1,1,74],"6":[5,1,254],"7":[1,5,39],"8":[3,1,30],"9":[3,3,144],"10":[3,3,177],"11":[3,3,68],"12":[4,3,109],"13":[2,3,29],"14":[1,1,100],"15":[10,1,128],"16":[1,1,21],"17":[5,1,34],"18":[4,5,168],"19":[2,5,255],"20":[1,5,84],"21":[4,1,147],"22":
[…]
