### 2024-11-13 EHTJulia/ScatteringOptics.jl#23 @kazuakiyama
@Edenhofer
Now we implemented unit tests and just released v0.1.5. We implemented two types of tests, which is called from the main `test/runtests.jl`.

- `test/internalconsistency.jl`: Internal self-consistency, which doesn't depend on the input models
- `test/simtest.jl`: end-to-end tests based on comparison with reference data produced by ourselves using `test/data/datagen.jl`

This seems to bring a very good codecov coverage of [94%](https://app.codecov.io/github/EHTJulia/ScatteringOptics.jl) for v0.1.5. Those unit tests will be run by GitHub CI workflow using the standard setting for Julia packages. In Julia, you can run the test by
```julia
using Pkg
Pkg.add("ScatteringOptics")
Pkg.update("ScatteringOptics")
Pkg.test("ScatteringOptics")
```
