### 2024-11-13 #6354 @kazuakiyama
@Edenhofer
unit tests are now implemented in the new version v0.1.5. See the issues here.
https://github.com/EHTJulia/ScatteringOptics.jl/issues/23

We are working on benchmark scripts now, and will keep you posted.
