### 2024-11-13 4fb6be1
seperate the installation guide from introduction
