+++ docs/src/installation.md
@@ -1,0 +1,22 @@
+
+# Installation
+The package is registered in the Julia standard repository, and installable through the standard ways by
+```julia
+using Pkg
+Pkg.add("ScatteringOptics")
+```
+or
+```julia
+]add ScatteringOptics
+```
+
+## Unit Tests
+The installed package can be tested in your local enviroment through the standard ways by
+```julia
+using Pkg
+Pkg.test("ScatteringOptics")
+```
+or
+```julia
+]test ScatteringOptics
+```
