### 2024-11-14 EHTJulia/ScatteringOptics.jl#23 @Edenhofer
The tests look great! Thank you for implementing this :)
