+++ v0.1.6/assets/introduction.md.C4siYZ6N.js
@@ -1,0 +1,1 @@
+import{_ as t,c as i,a5 as a,o as n}from"./chunks/framework.BI2mU9dR.js";const m=JSON.parse('{"title":"Introduction","description":"","frontmatter":{},"headers":[],"relativePath":"introduction.md","filePath":"introduction.md","lastUpdated":null}'),r={name:"introduction.md"};function o(s,e,l,c,d,h){return n(),i("div",null,e[0]||(e[0]=[a('<h1 id="introduction" tabindex="-1">Introduction <a class="header-anchor" href="#introduction" aria-label="Permalink to &quot;Introduction&quot;">​</a></h1><p><a href="https://github.com/EHTJulia/ScatteringOptics.jl" target="_blank" rel="noreferrer"><code>ScatteringOptics.jl</code></a> is an astronomy software package developed in the Julia programming language for interstellar scintillation observed at radio wavelengths. It implements physical models for the anisotropic scattering of radio waves, which arise from turbulence in the ionized interstellar medium.</p><p>This toolkit is aiming to provide capabilities of simulating and modeling the temporal, spatial, and spectral effects of interstellar scintillation primarily in the strong scattering regime, taking advantage of Julia&#39;s speed and composability. The package is designed to work with sky models and interferometric data types from the advanced Bayesian radio interferometric modeling package <a href="https://github.com/ptiede/Comrade.jl" target="_blank" rel="noreferrer"><code>Comrade.jl</code></a>.</p><p>The package provides essential functionalities for modeling, analyzing, and interpreting the images of the Galactic Center&#39;s supermassive black hole, Sagittarius A*, especially with the Event Horizon Telescope, and also the images of extremely high brightness temperature emissions in active galactic nuclei using space very long baseline interferometry.</p><h2 id="What-are-interstellar-scattering-and-scintillation?-Why-are-they-important?" tabindex="-1">What are interstellar scattering and scintillation? Why are they important? <a class="header-anchor" href="#What-are-interstellar-scattering-and-scintillation?-Why-are-they-important?" aria-label="Permalink to &quot;What are interstellar scattering and scintillation? Why are they important? {#What-are-interstellar-scattering-and-scintillation?-Why-are-they-important?}&quot;">​</a></h2><p>Scintillation is a well-known phenomenon in astronomy. Stars appear to twinkle in visible light due to the scattering of their light by the Earth&#39;s atmosphere. In radio wavelengths, fluctuations in the electron density of the ionized interstellar plasma cause scattering of radio waves, resulting in the scintillation of compact radio sources in the sky. These phenomena are referred to as <em>interstellar scattering</em> and <em>scintillation</em>, respectively. Interstellar scintillation produces temporal and spectral modulations (i.e., twinkling) in the brightness of objects, as well as distortion of their images.</p><p>Scattering has become increasingly important in high-angular-resolution studies of compact radio sources at micro-arcsecond scales using very long baseline interferometry (VLBI). Notable examples include the black hole imaging of the supermassive black hole, Sagittarius A* (Sgr A*), at the center of the Milky Way with the Event Horizon Telescope, and studies of extremely high brightness temperature emissions in active galactic nuclei (AGNs) using space VLBI, exemplified by projects like RadioAstron. In both scenarios, the observed interferometric data are influenced by both diffractive and refractive scattering effects.</p><p>For a more detailed description of the scattering models implemented in this package, see <a href="/ScatteringOptics.jl/v0.1.6/math#Brief-Introduction-to-Interstellar-Scattering">Brief Introduction to Interstellar Scattering</a>.</p>',8)]))}const g=t(r,[["render",o]]);export{m as __pageData,g as default};
