### 2024-11-14 dc6d870
add NPZ dependency
