### 2024-11-22 ae9b810
first commit of some benchmark scripts
