+++ examples/benchmarks/1_scattering_kernel/compute_kernel.jl
@@ -1,5 +1,6 @@
 using Pkg
 Pkg.activate(dirname(@__DIR__))
+Pkg.instantiate()
 
 using CairoMakie
 using CSV
