+++ examples/benchmarks/0_model_initializaiton/model_init.jl
@@ -1,5 +1,6 @@
 using Pkg
 Pkg.activate(dirname(@__DIR__))
+Pkg.instantiate()
 
 using BenchmarkTools
 using CSV
