### 2024-11-22 743490a
produce outputs of benchmarks 0 and 1, add setup scripts
