+++ examples/benchmarks/Project.toml
@@ -3,9 +3,9 @@ BenchmarkTools = "6e4b80f9-dd63-53aa-95a3-0cdb28fa8baf"
 CSV = "336ed68f-0bac-5ca0-87d4-7b16caf5d00b"
 CairoMakie = "13f3f980-e62b-5c42-98c6-ff1f3baf88f0"
 DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
+LaTeXStrings = "b964fa9f-0449-5b57-a5c2-d3ea65f4040f"
 PrettyTables = "08abe8d2-0d0c-5749-adfa-8a2ac140af0d"
 ProgressBars = "49802e3a-d2f1-5c88-81d8-b72133a6f568"
-Pyehtim = "3d61700d-6e5b-419a-8e22-9c066cf00468"
 PythonCall = "6099a3de-0909-46bc-b1f4-468b9a2dfc0d"
 ScatteringOptics = "e317c96f-4a6f-4ae9-8016-294fb9063ff5"
 StableRNGs = "860ef19b-820b-49d6-a774-d7a799459cd3"

[…]
