### 2024-11-23 d5d1757
add full scattering simulations
