### 2024-11-23 ef08987
add the third test for the kernel consistency
