+++ docs/src/benchmarks.md
@@ -4,39 +4,91 @@ CurrentModule = ScatteringOptics
 # Benchmarks
 The Julia implementation of the EHT scattering framework takes advantage of Julia's fast Auto-Differentiation (AD) capabilities, which accelerate the simulation of interstellar scattering. As the overall imaging process can consist of thousands to tens of thousands of iterations in order to optimize scattering parameters and the final output image, even a small imprdovements in simulation speed can result in significant cuts in runtime. It is then crucial that our implementation performs faster than the existing python framework, `eht-imaging`, and produces consistent results. Here, we present benchmarks between the two implementations.
 
-## Speed
+All the examples are in `examples/benchmarks` directories. In below, benchmark results were obtained under the following enviroment.
+- Ubuntu 22.04 LTS
+- Intel(R) Core(TM) i7-5820K CPU @ 3.30GHz
+- Julia Version 1.10.6
+- ScatteringOptics v0.1.6
+- Python 3.11 (installed through mamba/miniforge3)
+- eht-imaging v1.2.8
 
-With Julia's speed improvements, the `ScatteringOptics.jl` scattering model loads 100 times faster than the `eht-imaging`'s scattering model, which both require to semi-analytically solve co-efficients and precompute the phase structure function for the corresponding scattering parameters. 
+You can set up your own enviroment by
+```bash
+# this command will use mamba command. You can instead use conda command by editing the script.
+bash setup_python.bash
+```
+or alternatively you can reproduce the enviroment tested by
+```bash
+# you can replace mamba with conda
+mamba install --name py311_sobenchmark --file conda_list.txt
 
-The scattering kernel may be called upon to compute visibilities for a given sample of Fourier space points (see [Simulate Diffractive Scattering](@ref) for the tutorial), which is important to mitigate the diffractive scattering in the sky model (e.g. image) reconstruction. In these computations, `ScatteringOptics.jl` also exhibits speed improvements up to 100 times over the three included models (Dipole, Boxcar, and Von Mises; see [Use Non-default Models](@ref) for more info) and varying observing wavelengths. These computation runtimes are compared over a range of sample in the plots below. 
+# or alternatively
+mamba install --name py311_sobenchmark --file conta_explicit_env.txt
+```
+
+Then you can run all the tests by
+```bash
+bash run_benchmark.bash
+```
+It will take about an hour.
+
+## Model initialization
+Both `eht-imaging` and `ScatteringOptics.jl` solves semi-analytic equations to precompute constants (e.g. concentration parameter) required for the phase structure function and other related functions, for the given input scattering parameters. 
+For the future implementation of the joint inference for the scattering parameters, it is important that the model initialization is fast enough for a large number of iterations.
+
+
+| Package              | Dipole (Default)  |   von Mises       | Periodic BoxCar     | 
+| -------------------- | ----------------  | ----------------  | ------------------  |
+| ScatteringOptics.jl  |  0.26 +/- 0.15 ms |  0.17 +/- 0.08 ms |   0.51 +/- 0.11 ms  |
+| eht-imaging (Python) | 22.86 +/- 1.53 ms | 15.99 +/- 1.58 ms | 149.24 +/- 6.79 ms  |
+
+This shows `ScatteringOptics.jl` gives about 90-100 times speed boost for Dipole and von Mises models, and 300 times speed boost for the Periodic Boxcar model.
+The benchmark scripts are available in `examples/0_model_initialization`, and can be executed with
+```bash
+# run both python and julia
+bash run.bash
+```
+
+With this significant speed up, the Julia implementation paves the pathway for the joint modeling of scattering parameters and reconstructed images in self-consistent way. 
+
+## Scattering kernel computations
[…]
