+++ dev/assets/benchmarks.md._jcH4_k7.lean.js
@@ -1,1 +1,0 @@
-import{_ as p,c as i,a5 as l,G as a,w as s,j as t,a as o,B as m,o as r}from"./chunks/framework.BMbWnoF-.js";const c="/ScatteringOptics.jl/dev/assets/speed_dipole.0zob2s6s.png",h="/ScatteringOptics.jl/dev/assets/speed_boxcar.DNaYty9n.png",u="/ScatteringOptics.jl/dev/assets/speed_vonmises.QMxRx1fN.png",g="/ScatteringOptics.jl/dev/assets/kernel87.aE8RPnav.png",T="/ScatteringOptics.jl/dev/assets/kernel36.BxEMma2y.png",S=JSON.parse('{"title":"Benchmarks","description":"","frontmatter":{},"headers":[],"relativePath":"benchmarks.md","filePath":"benchmarks.md","lastUpdated":null}'),Q={name:"benchmarks.md"},f={class:"MathJax",jax:"SVG",style:{direction:"ltr",position:"relative"}},x={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-0.05ex"},xmlns:"http://www.w3.org/2000/svg",width:"4.495ex",height:"2.005ex",role:"img",focusable:"false",viewBox:"0 -864 1986.7 886","aria-hidden":"true"},w={class:"MathJax",jax:"SVG",style:{direction:"ltr",position:"relative"}},b={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-0.05ex"},xmlns:"http://www.w3.org/2000/svg",width:"4.495ex",height:"2.02ex",role:"img",focusable:"false",viewBox:"0 -871.1 1986.7 893.1","aria-hidden":"true"};function v(k,e,y,M,H,V){const n=m("PluginTabsTab"),d=m("PluginTabs");return r(),i("div",null,[e[14]||(e[14]=l('<h1 id="benchmarks" tabindex="-1">Benchmarks <a class="header-anchor" href="#benchmarks" aria-label="Permalink to &quot;Benchmarks&quot;">​</a></h1><p>The Julia implementation of the EHT scattering framework takes advantage of Julia&#39;s fast Auto-Differentiation (AD) capabilities, which accelerate the simulation of interstellar scattering. As the overall imaging process can consist of thousands to tens of thousands of iterations in order to optimize scattering parameters and the final output image, even a small imprdovements in simulation speed can result in significant cuts in runtime. It is then crucial that our implementation performs faster than the existing python framework, <code>eht-imaging</code>, and produces consistent results. Here, we present benchmarks between the two implementations.</p><h2 id="speed" tabindex="-1">Speed <a class="header-anchor" href="#speed" aria-label="Permalink to &quot;Speed&quot;">​</a></h2><p>With Julia&#39;s speed improvements, the <code>ScatteringOptics.jl</code> scattering model loads 100 times faster than the <code>eht-imaging</code>&#39;s scattering model, which both require to semi-analytically solve co-efficients and precompute the phase structure function for the corresponding scattering parameters.</p><p>The scattering kernel may be called upon to compute visibilities for a given sample of Fourier space points (see <a href="/ScatteringOptics.jl/dev/diffractive#Simulate-Diffractive-Scattering">Simulate Diffractive Scattering</a> for the tutorial), which is important to mitigate the diff
[…]

[…]
