### 2024-11-23 #6354 @kazuakiyama
@Edenhofer
@annatartaglia and I have released the new version `v0.1.7`. In the latest version, we put benchmark scripts in `examples/benchmarks`. Those scripts include two more new speed tests. Please have a look at the updated documentation page below for the usage of scripts and results obtained in a local environment.
https://ehtjulia.github.io/ScatteringOptics.jl/v0.1.7/benchmarks
