### 2024-11-26 #6354 @Edenhofer
The benchmarks look good. Thank you for including them! One minor comment: I would suggest rewording the following
> exhibits speed improvements up to 100 times over the three included models
This does not seem like a fair interpretation of the figure which shows a speedup by a factor of ~30 only for the very conservative example of 10 points.
