### 2024-11-27 798f0b5
amend some statements based on the comments from the on-going JOSS review.
