+++ dev/assets/benchmarks.md.CwIsUe_B.js
@@ -4,7 +4,7 @@ import{_ as d,c as s,a5 as a,G as l,w as o,j as t,a as i,B as r,o as n}from"./ch
 <span class="line"></span>
 <span class="line"><span style="--shiki-light:#6A737D;--shiki-dark:#6A737D;"># or alternatively</span></span>
 <span class="line"><span style="--shiki-light:#6F42C1;--shiki-dark:#B392F0;">mamba</span><span style="--shiki-light:#032F62;--shiki-dark:#9ECBFF;"> install</span><span style="--shiki-light:#005CC5;--shiki-dark:#79B8FF;"> --name</span><span style="--shiki-light:#032F62;--shiki-dark:#9ECBFF;"> py311_sobenchmark</span><span style="--shiki-light:#005CC5;--shiki-dark:#79B8FF;"> --file</span><span style="--shiki-light:#032F62;--shiki-dark:#9ECBFF;"> conta_explicit_env.txt</span></span></code></pre></div><p>Then you can run all the tests by</p><div class="language-bash vp-adaptive-theme"><button title="Copy Code" class="copy"></button><span class="lang">bash</span><pre class="shiki shiki-themes github-light github-dark vp-code" tabindex="0"><code><span class="line"><span style="--shiki-light:#6F42C1;--shiki-dark:#B392F0;">bash</span><span style="--shiki-light:#032F62;--shiki-dark:#9ECBFF;"> run_benchmark.bash</span></span></code></pre></div><p>It will take about an hour.</p><h2 id="Model-initialization" tabindex="-1">Model initialization <a class="header-anchor" href="#Model-initialization" aria-label="Permalink to &quot;Model initialization {#Model-initialization}&quot;">​</a></h2><p>Both <code>eht-imaging</code> and <code>ScatteringOptics.jl</code> solves semi-analytic equations to precompute constants (e.g. concentration parameter) required for the phase structure function and other related functions, for the given input scattering parameters. For the future implementation of the joint inference for the scattering parameters, it is important that the model initialization is fast enough for a large number of iterations.</p><table tabindex="0"><thead><tr><th style="text-align:right;">Package</th><th style="text-align:right;">Dipole (Default)</th><th style="text-align:right;">von Mises</th><th style="text-align:right;">Periodic BoxCar</th></tr></thead><tbody><tr><td style="text-align:right;">ScatteringOptics.jl</td><td style="text-align:right;">0.26 +/- 0.15 ms</td><td style="text-align:right;">0.17 +/- 0.08 ms</td><td style="text-align:right;">0.51 +/- 0.11 ms</td></tr><tr><td style="text-align:right;">eht-imaging (Python)</td><td style="text-align:right;">22.86 +/- 1.53 ms</td><td style="text-align:right;">15.99 +/- 1.58 ms</td><td style="text-align:right;">149.24 +/- 6.79 ms</td></tr></tbody></table><p>This shows <code>ScatteringOptics.jl</code> gives about 90-100 times speed boost for Dipole and von Mises models, and 300 times speed boost for the Periodic Boxcar model. The benchmark scripts are available in <code>examples/0_model_initialization</code>, and can be executed with</p><div class="language-bash vp-adaptive-theme"><button title="Copy Code" class="copy"></button><span class="lang">bash</span><pre class="shiki shiki-themes github-light github-dark vp-code" tabindex="0"><code><span class="line"><span style="--shiki-light:#6A737D;--shiki-dark:#6A737D;"># run both python and julia</span></span>
[…]

[…]
