+++ v0.1.8/assets/custommodels.md.sGwDZR62.lean.js
@@ -1,0 +1,94 @@
+import{_ as e,c as t,a5 as n,j as i,a,o as l}from"./chunks/framework.ClVBwEoY.js";const D=JSON.parse('{"title":"Define Your Own Scattering Model","description":"","frontmatter":{},"headers":[],"relativePath":"custommodels.md","filePath":"custommodels.md","lastUpdated":null}'),h={name:"custommodels.md"},p={class:"MathJax",jax:"SVG",style:{direction:"ltr",position:"relative"}},k={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-0.666ex"},xmlns:"http://www.w3.org/2000/svg",width:"2.12ex",height:"2.236ex",role:"img",focusable:"false",viewBox:"0 -694 937 988.2","aria-hidden":"true"},r={class:"MathJax",jax:"SVG",style:{direction:"ltr",position:"relative"}},d={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-0.666ex"},xmlns:"http://www.w3.org/2000/svg",width:"2.12ex",height:"2.236ex",role:"img",focusable:"false",viewBox:"0 -694 937 988.2","aria-hidden":"true"};function E(g,s,o,c,y,m){return l(),t("div",null,[s[11]||(s[11]=n(`<h1 id="Define-Your-Own-Scattering-Model" tabindex="-1">Define Your Own Scattering Model <a class="header-anchor" href="#Define-Your-Own-Scattering-Model" aria-label="Permalink to &quot;Define Your Own Scattering Model {#Define-Your-Own-Scattering-Model}&quot;">​</a></h1><p>While this package covers all three probabilistic models published in literature as of 2024 (see <a href="/ScatteringOptics.jl/v0.1.8/math#Brief-Introduction-to-Interstellar-Scattering">Brief Introduction to Interstellar Scattering</a>), users may define a custom probabilistic model leveledging a high level abstraction of data types implemented in this package.</p><p>To make your own custom model, you only need to define a custom type of <a href="/ScatteringOptics.jl/v0.1.8/api#ScatteringOptics.AbstractScatteringModel">ScatteringOptics.AbstractScatteringModel</a>. You need to define the constructor of the type with the standardized parameters set in the literature. Here is a quick example. As long as your model is compatible with the framework of the semianalytic models in Psaltis et al. [1], you need to change only few lines.</p><div class="language-julia vp-adaptive-theme"><button title="Copy Code" class="copy"></button><span class="lang">julia</span><pre class="shiki shiki-themes github-light github-dark vp-code" tabindex="0"><code><span class="line"><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">using</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;"> ScatteringOptics</span></span>
+<span class="line"></span>
[…]
