### 2024-11-27 #6354 @kazuakiyama
This was indeed a remnant from the previous version based on an earlier benchmark. Now this has been corrected for the latest version. Thanks for pointing this out!
https://ehtjulia.github.io/ScatteringOptics.jl/v0.1.8/benchmarks
