+++ dev/assets/chunks/@localSearchIndexroot.3lh2vDzj.js
@@ -1,1 +1,0 @@
-const e='{"documentCount":31,"nextId":31,"documentIds":{"0":"/ScatteringOptics.jl/dev/api#ScateringOptics.jl-API","1":"/ScatteringOptics.jl/dev/api#index","2":"/ScatteringOptics.jl/dev/api#docstrings","3":"/ScatteringOptics.jl/dev/benchmarks#benchmarks","4":"/ScatteringOptics.jl/dev/benchmarks#Model-initialization","5":"/ScatteringOptics.jl/dev/benchmarks#Scattering-kernel-computations","6":"/ScatteringOptics.jl/dev/benchmarks#End-to-end-scattering-simulations","7":"/ScatteringOptics.jl/dev/benchmarks#Consistency-of-the-scattering-kernel","8":"/ScatteringOptics.jl/dev/custommodels#Define-Your-Own-Scattering-Model","9":"/ScatteringOptics.jl/dev/custommodels#references","10":"/ScatteringOptics.jl/dev/diffractive#Simulate-Diffractive-Scattering","11":"/ScatteringOptics.jl/dev/diffractive#Loading-your-image","12":"/ScatteringOptics.jl/dev/diffractive#Simulate-diffractive-scattering","13":"/ScatteringOptics.jl/dev/diffractive#A-quick-shortcut","14":"/ScatteringOptics.jl/dev/diffractive#Save-the-tutorial-data","15":"/ScatteringOptics.jl/dev/diffractive#Unit-Tests","16":"/ScatteringOptics.jl/dev/installation#installation","17":"/ScatteringOptics.jl/dev/installation#Unit-Tests","18":"/ScatteringOptics.jl/dev/introduction#introduction","19":"/ScatteringOptics.jl/dev/introduction#What-are-interstellar-scattering-and-scintillation?-Why-are-they-important?","20":"/ScatteringOptics.jl/dev/math#Brief-Introduction-to-Interstellar-Scattering","21":"/ScatteringOptics.jl/dev/math#Scattering-Regimes-in-Scope","22":"/ScatteringOptics.jl/dev/math#Scattering-Model","23":"/ScatteringOptics.jl/dev/math#references","24":"/ScatteringOptics.jl/dev/nondefaultmodels#Use-Non-default-Models","25":"/ScatteringOptics.jl/dev/nondefaultmodels#references","26":"/ScatteringOptics.jl/dev/refractive#Simulate-Refractive-Scattering","27":"/ScatteringOptics.jl/dev/refractive#Loading-your-image","28":"/ScatteringOptics.jl/dev/refractive#Simulating-Refractive-Scattering","29":"/ScatteringOptics.jl/dev/refractive#Save-the-tutorial-data-and-do-some-unit-tests","30":"/ScatteringOptics.jl/dev/refractive#Unit-Tests"},"fieldIds":{"title":0,"titles":1,"text":2},"fieldLength":{"0":[3,1,1],"1":[1,3,42],"2":[1,3,478],"3":[1,1,153],"4":[2,1,117],"5":[3,1,153],"6":[5,1,85],"7":[5,1,85],"8":[5,1,254],"9":[1,5,39],"10":[3,1,30],"11":[3,3,144],"12":[3,3,177],"13":
[…]
