### 2024-12-12 #6354 @tomkimpson
All my comments have been addressed. I am pleased with the [new unit tests](https://github.com/EHTJulia/ScatteringOptics.jl/issues/23) and the updates to the tutorial. I am happy for publication. Well done! 🚀
