### 2024-12-19 b2c38a6
Paper edits for JOSS review.
