+++ paper/paper.bib
@@ -1,7 +1,7 @@
 @ARTICLE{Johnson_2015,
        author = {{Johnson}, Michael D. and {Gwinn}, Carl R.},
         title = "{Theory and Simulations of Refractive Substructure in Resolved Scatter-broadened Images}",
-      journal = {\apj},
+      journal = {Astrophysical Journal},
      keywords = {ISM: structure, Galaxy: nucleus, radio continuum: ISM, scattering, techniques: interferometric, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Astrophysics of Galaxies},
          year = 2015,
         month = jun,
@@ -20,7 +20,7 @@ archivePrefix = {arXiv},
 @article{Johnson_2016,
        author = {{Johnson}, Michael D.},
         title = "{Stochastic Optics: A Scattering Mitigation Framework for Radio Interferometric Imaging}",
-      journal = {\apj},
+      journal = {Astrophysical Journal},
      keywords = {ISM: structure, radio continuum: ISM, scattering, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Astrophysics of Galaxies},
          year = 2016,
         month = dec,
@@ -39,7 +39,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Johnson_2016_3C273,
        author = {{Johnson}, Michael D. and {Kovalev}, Yuri Y. and {Gwinn}, Carl R. and {Gurvits}, Leonid I. and {Narayan}, Ramesh and {Macquart}, Jean-Pierre and {Jauncey}, David L. and {Voitsik}, Peter A. and {Anderson}, James M. and {Sokolovsky}, Kirill V. and {Lisakov}, Mikhail M.},
         title = "{Extreme Brightness Temperatures and Refractive Substructure in 3C273 with RadioAstron}",
-      journal = {\apjl},
+      journal = {Astrophysical Journal, Letters},
      keywords = {ISM: structure, quasars: individual: 3C 273, scattering, techniques: high angular resolution, techniques: interferometric, turbulence, Astrophysics - High Energy Astrophysical Phenomena, Astrophysics - Instrumentation and Methods for Astrophysics},
          year = 2016,
         month = mar,
@@ -58,7 +58,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Johnson_Narayan_2016,
        author = {{Johnson}, Michael D. and {Narayan}, Ramesh},
         title = "{The Optics of Refractive Substructure}",
-      journal = {\apj},
+      journal = {Astrophysical Journal},
      keywords = {Galaxy: nucleus, ISM: structure, radio continuum: ISM, scattering, techniques: interferometric, turbulence, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Astrophysics of Galaxies},
          year = 2016,
         month = aug,
@@ -91,7 +91,7 @@ journal = {The Astrophysical Journal},
 @ARTICLE{EHTSgrAPaper1,
        author = {{Event Horizon Telescope Collaboration et al.}},
         title = "{First Sagittarius A* Event Horizon Telescope Results. I. The Shadow of the Supermassive Black Hole in the Center of the Milky Way}",
-      journal = {\apjl},
+      journal = {Astrophysical Journal, Letters},
      keywords = {Black holes, Kerr black holes, Rotating black holes, Heterodyne interferometry, Galactic center, 162, 886, 1406, 726, 565},
          year = 2022,
         month = may,
@@ -107,7 +107,7 @@ journal = {The Astrophysical Journal},
 @ARTICLE{EHTSgrAPaper2,
        author = {{Event Horizon Telescope Collaboration et al.}},
[…]
