### 2024-12-20 659ccbf
Update affiliations in paper.md
