### 2024-12-23 ef093e8
updating the software version and dependencies
