+++ docs/make.jl
@@ -2,25 +2,20 @@ using ScatteringOptics
 using Documenter
 using DocumenterVitepress
 
-DocMeta.setdocmeta!(ScatteringOptics, :DocTestSetup, :(using ScatteringOptics); recursive=true)
+DocMeta.setdocmeta!(
+    ScatteringOptics, :DocTestSetup, :(using ScatteringOptics); recursive=true
+)
 
 # Tutorial files
 TUTORIALS = [
-    "Getting Started" => [
-        "installation.md",
-        "diffractive.md",
-        "refractive.md",
-    ],
-    "Advanced" =>[
-        "nondefaultmodels.md",
-        "custommodels.md",
-    ],
+    "Getting Started" => ["installation.md", "diffractive.md", "refractive.md"],
+    "Advanced" => ["nondefaultmodels.md", "custommodels.md"],
 ]
 
 # Formatter
-format = MarkdownVitepress(
-    repo = "github.com/EHTJulia/ScatteringOptics.jl",
-    clean_md_output = false,    #  this needs to be false for the manual data copy
+format = MarkdownVitepress(;
+    repo="github.com/EHTJulia/ScatteringOptics.jl",
+    clean_md_output=false,    #  this needs to be false for the manual data copy
 )
 
 # Keywords for Documenter.jl
@@ -29,7 +24,7 @@ makedocs_kwargs = (
     authors="Anna Tartaglia, Kazunori Akiyama",
     repo="github.com/EHTJulia/ScatteringOptics.jl/blob/{commit}{path}#{line}",
     sitename="ScatteringOptics.jl",
-    format = format,
+    format=format,
     pages=[
         "Home" => "index.md",
         "Introduction" => "introduction.md",
@@ -49,13 +44,14 @@ makedocs(; makedocs_kwargs...)
 
 # shortcut to some directories
 doc = Documenter.Document(; makedocs_kwargs...) # snipet from makedocs.jl in Documenter.jl
-builddir = isabspath(doc.user.build) ? doc.user.build : joinpath(doc.user.root, doc.user.build) 
+builddir =
+    isabspath(doc.user.build) ? doc.user.build : joinpath(doc.user.root, doc.user.build)
 mddir = joinpath(builddir, format.md_output_path)
 fsdir = joinpath(builddir, "final_site")
 
 # manually copy data files into final_site as DocumenterVitepress will ignore files in src/data.
 @info "make.jl: manually copy files in src/data"
-mkpath(joinpath(fsdir,"data"))
+mkpath(joinpath(fsdir, "data"))
 contents = readdir(joinpath(mddir, "data"))
 for item in contents
     src = joinpath(mddir, "data", item)
@@ -68,9 +64,9 @@ if format.deploy_decision === nothing
     deploy_config = Documenter.auto_detect_deploy_system()
     deploy_decision = Documenter.deploy_folder(
         deploy_config;
-        repo = format.repo, # this must be the full URL!
-        devbranch = format.devbranch,
-        devurl = format.devurl,
+        repo=format.repo, # this must be the full URL!
+        devbranch=format.devbranch,
+        devurl=format.devurl,
         push_preview=true,
     )
 else
@@ -80,15 +76,15 @@ end
 # manually clean up the build directory if the documentation to be deployed.
 if deploy_decision.all_ok
     @info "make.jl: manually cleaning up Markdown output."
-    rm(mddir; recursive = true)
-    
+    rm(mddir; recursive=true)
+
     contents = readdir(fsdir)
     for item in contents
         src = joinpath(fsdir, item)
         dst = joinpath(builddir, item)
         cp(src, dst)
     end
-    rm(fsdir; recursive = true)
[…]
