### 2025-01-16 #6354 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
