### 2025-01-16 #6354 @dfm
@Edenhofer, @tomkimpson — Many thanks for your reviews here! JOSS relies upon the volunteer effort of people like you and we simply wouldn't be able to do this without you!!

@annatartaglia — Your paper has now been handed off to the track EiC who may have some final edits, before handling the final processing.
