### 2025-01-16 #6354 @warrickball
I'm having a first look at this now. @tomkimpson, on the basis of [this comment of yours, above](https://github.com/openjournals/joss-reviews/issues/6354#issuecomment-2537861512), I've gone ahead and ticked off everything in [your review checklist](https://github.com/openjournals/joss-reviews/issues/6354#issuecomment-1970299495).

@annatartaglia, there a few things to address in the current metadata:
1. Could you add countries to the affiliations?
2. The version listed in the Zenodo sidebar is v1, whereas the reviewed version (and the version in the tarball) is v0.1.9. Could you fix the Zenodo version?
3. Could you correct the license in the Zenodo entry to MIT License, as in the code repo?
4. I'll take a closer look at the references when I proofread the article but could you update the Julia reference to [the one  linked to](https://github.com/JuliaLang/julia/blob/master/CITATION.bib) in [their documentation](https://julialang.org/research/)?

I intend to give the paper a closer look in the next few days, after which I'll probably propose some changes in a pull request.
