### 2025-01-17 f579200
Change Julia citation in bibliography
