+++ paper/paper.bib
@@ -173,21 +173,17 @@ journal = {The Astrophysical Journal},
       primaryClass={astro-ph.HE}
 }
 
-@ARTICLE{Bezanson_2012,
-       author = {{Bezanson}, Jeff and {Karpinski}, Stefan and {Shah}, Viral B. and {Edelman}, Alan},
-        title = "{Julia: A Fast Dynamic Language for Technical Computing}",
-      journal = {arXiv e-prints},
-     keywords = {Computer Science - Programming Languages, Computer Science - Computational Engineering, Finance, and Science, D.3.2},
-         year = 2012,
-        month = sep,
-          eid = {arXiv:1209.5145},
-        pages = {arXiv:1209.5145},
-          doi = {10.48550/arXiv.1209.5145},
-archivePrefix = {arXiv},
-       eprint = {1209.5145},
- primaryClass = {cs.PL},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2012arXiv1209.5145B},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+@article{Julia_2017,
+    title={Julia: A fresh approach to numerical computing},
+    author={Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B},
+    journal={SIAM {R}eview},
+    volume={59},
+    number={1},
+    pages={65--98},
+    year={2017},
+    publisher={SIAM},
+    doi={10.1137/141000671},
+    url={https://epubs.siam.org/doi/10.1137/141000671}
 }
 
 @ARTICLE{Rickett_1990,
