### 2025-01-17 50ddc5b
Add countries to affiliations, update Julia citation
