+++ dev/assets/index.md.8y_-OsQu.js
@@ -1,1 +1,0 @@
-import{_ as t,c as e,o as a}from"./chunks/framework.Cw7Jlqmt.js";const p=JSON.parse('{"title":"","description":"","frontmatter":{"layout":"home","hero":{"name":"ScatteringOptics.jl","tagline":"A Julia package to model and simulate interstellar scattering and scintillation","actions":[{"theme":"brand","text":"Introduction","link":"/introduction"},{"theme":"alt","text":"Tutorials","link":"/installation"},{"theme":"alt","icon":"github","text":"View on Github","link":"https://github.com/EHTJulia/ScatteringOptics.jl"}]}},"headers":[],"relativePath":"index.md","filePath":"index.md","lastUpdated":null}'),n={name:"index.md"};function i(o,l,r,c,s,d){return a(),e("div")}const u=t(n,[["render",i]]);export{p as __pageData,u as default};
