+++ paper/paper.bib
@@ -172,7 +172,6 @@ journal = {The Astrophysical Journal},
       archivePrefix={arXiv},
       primaryClass={astro-ph.HE}
 }
-
 @article{Julia_2017,
     title={Julia: A fresh approach to numerical computing},
     author={Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B},
