### 2025-01-17 #6354 @annatartaglia
Hi @warrickball, I've made all the requested amendments to the paper and Zenodo. Let me know if it looks good, or if more changes are needed.

Additionally, let me know if we should make a patch release (v0.1.10) to the repository and update the Zenodo accordingly. Thanks!
