### 2025-01-24 #6354 @warrickball
Hi @annatartaglia, I've proposed what I hope are some final changes in a [pull request](https://github.com/EHTJulia/ScatteringOptics.jl/pull/32). Once those are merged (presuming they're all okay), I'll give the publication pipeline another test run to check the relevant metadata (e.g. author affiliations) is also correctly updated.

All the archive metadata looks good now. Changes to the paper don't require a new archive.
