+++ dev/assets/benchmarks.md.BODcNCHr.js
@@ -1,4 +1,4 @@
-import{_ as d,c as s,a5 as a,G as l,w as o,j as t,a as i,B as r,o as n}from"./chunks/framework.DgaFFy6w.js";const c="/ScatteringOptics.jl/dev/assets/bench_kernel_speed.CfCP8P3k.png",m="/ScatteringOptics.jl/dev/assets/bench_full_scattering.CQTmu7Cr.png",g="/ScatteringOptics.jl/dev/assets/bench_kernel_accuracy.DEf3022g.png",B=JSON.parse('{"title":"Benchmarks","description":"","frontmatter":{},"headers":[],"relativePath":"benchmarks.md","filePath":"benchmarks.md","lastUpdated":null}'),u={name:"benchmarks.md"},k={class:"MathJax",jax:"SVG",style:{direction:"ltr",position:"relative"}},b={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-0.05ex"},xmlns:"http://www.w3.org/2000/svg",width:"4.495ex",height:"2.005ex",role:"img",focusable:"false",viewBox:"0 -864 1986.7 886","aria-hidden":"true"},f={class:"MathJax",jax:"SVG",style:{direction:"ltr",position:"relative"}},y={style:{overflow:"visible","min-height":"1px","min-width":"1px","vertical-align":"-0.05ex"},xmlns:"http://www.w3.org/2000/svg",width:"4.495ex",height:"2.02ex",role:"img",focusable:"false",viewBox:"0 -871.1 1986.7 893.1","aria-hidden":"true"};function v(x,e,T,w,Q,F){const h=r("PluginTabsTab"),p=r("PluginTabs");return n(),s("div",null,[e[10]||(e[10]=a(`<h1 id="benchmarks" tabindex="-1">Benchmarks <a class="header-anchor" href="#benchmarks" aria-label="Permalink to &quot;Benchmarks&quot;">​</a></h1><p>The Julia implementation of the EHT scattering framework takes advantage of Julia&#39;s fast Auto-Differentiation (AD) capabilities, which accelerate the simulation of interstellar scattering. As the overall imaging process can consist of thousands to tens of thousands of iterations in order to optimize scattering parameters and the final output image, even a small improvements in simulation speed can result in significant cuts in runtime. It is then crucial that our implementation performs faster than the existing python framework, <code>eht-imaging</code>, and produces consistent results. Here, we present benchmarks between the two implementations.</p><p>All the examples are in <code>examples/benchmarks</code> directories. In below, benchmark results were obtained under the following enviroment.</p><ul><li><p>Ubuntu 22.04 LTS</p></li><li><p>Intel(R) Core(TM) i7-5820K CPU @ 3.30GHz</p></li><li><p>Julia Version 1.10.6</p></li><li><p>ScatteringOptics v0.1.6</p></li><li><p>Python 3.11 (installed through mamba/miniforge3)</p></li><li><p>eht-imaging v1.2.8</p></li></ul><p>You can set up your own enviroment by</p><div class="language-bash vp-adaptive-theme"><button title="Copy Code" class="copy"></button><span class="lang">bash</span><pre class="shiki shiki-themes github-light github-dark vp-code" tabindex="0"><code><span class="line"><span style="--shiki-light:#6A737D;--shiki-dark:#6A737D;"># this command will use mamba command. You can instead use conda command by editing the script.</span></span>
[…]
