### 2025-01-26 #6354 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
