### 2025-01-26 13cce2e
Update README.md
Add a link to the JOSS paper.
