+++ README.md
@@ -1,7 +1,7 @@
 # ScatteringOptics.jl
-| Documentation | Code Status | Guidelines |
-| :------- | :------ | :------ |
-| [![Stable](https://img.shields.io/badge/docs-stable-blue.svg)](https://EHTJulia.github.io/ScatteringOptics.jl/stable/)[![Dev](https://img.shields.io/badge/docs-dev-blue.svg)](https://EHTJulia.github.io/ScatteringOptics.jl/dev/) | [![Build Status](https://github.com/EHTJulia/ScatteringOptics.jl/actions/workflows/CI.yml/badge.svg?branch=main)](https://github.com/EHTJulia/ScatteringOptics.jl/actions/workflows/CI.yml?query=branch%3Amain)[![codecov](https://codecov.io/github/EHTJulia/ScatteringOptics.jl/graph/badge.svg?token=OAARUNXSBY)](https://codecov.io/github/EHTJulia/ScatteringOptics.jl) | [![Code Style: Blue](https://img.shields.io/badge/code%20style-blue-4495d1.svg)](https://github.com/invenia/BlueStyle)[![ColPrac: Contributor's Guide on Collaborative Practices for Community Packages](https://img.shields.io/badge/ColPrac-Contributor's%20Guide-blueviolet)](https://github.com/SciML/ColPrac) |
+| Documentation | Paper | Code Status | Guidelines |
+| :------- | :------ | :------ | :------ |
+| [![Stable](https://img.shields.io/badge/docs-stable-blue.svg)](https://EHTJulia.github.io/ScatteringOptics.jl/stable/)[![Dev](https://img.shields.io/badge/docs-dev-blue.svg)](https://EHTJulia.github.io/ScatteringOptics.jl/dev/) | [![status](https://joss.theoj.org/papers/80e98109e1499d2f0f8e15838a3b272b/status.svg)](https://joss.theoj.org/papers/80e98109e1499d2f0f8e15838a3b272b) | [![Build Status](https://github.com/EHTJulia/ScatteringOptics.jl/actions/workflows/CI.yml/badge.svg?branch=main)](https://github.com/EHTJulia/ScatteringOptics.jl/actions/workflows/CI.yml?query=branch%3Amain)[![codecov](https://codecov.io/github/EHTJulia/ScatteringOptics.jl/graph/badge.svg?token=OAARUNXSBY)](https://codecov.io/github/EHTJulia/ScatteringOptics.jl) | [![Code Style: Blue](https://img.shields.io/badge/code%20style-blue-4495d1.svg)](https://github.com/invenia/BlueStyle)[![ColPrac: Contributor's Guide on Collaborative Practices for Community Packages](https://img.shields.io/badge/ColPrac-Contributor's%20Guide-blueviolet)](https://github.com/SciML/ColPrac) |
 
 `ScatteringOptics.jl` is an astronomy software package developed in the Julia programming language for interstellar scintillation observed at radio wavelengths. It implements physical models for the anisotropic scattering of radio waves, which arise from turbulence in the ionized interstellar medium. 
 
