### 2025-01-26 #6354 @kazuakiyama
Fantastic! Thank you very much @dfm for managing the entire process as the editor, @Edenhofer & @tomkimpson for constructive referees, and @warrickball for careful proofing! I am very excited to have the first JOSS paper for @annatartaglia and myself!
