### 2025-01-27 #6354 @warrickball
I'll echo @kazuakiyama's thanks to @Edenhofer & @tomkimpson for reviewing and @dfm for editing! JOSS simply wouldn't be possible without its community of volunteers.

Congratulations @annatartaglia & @kazuakiyama, your paper has been published in JOSS!
